% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassMethod-XGR.r
\name{aOnto}
\alias{aOnto}
\alias{print.aOnto}
\title{Definition for S3 class \code{aOnto}}
\usage{
aOnto(g, anno)

\method{print}{aOnto}(x, ...)
}
\arguments{
\item{g}{an igraph object}

\item{anno}{a list}

\item{x}{an object of class \code{aOnto}}

\item{...}{other parameters}
}
\value{
an object of S3 class \code{aOnto}
}
\description{
\code{aOnto} has 2 components: g, anno.
}
\examples{
\dontrun{
# Load the library
library(XGR)
}
RData.location <- "http://galahad.well.ox.ac.uk/bigdata"
\dontrun{
aOnto(g, anno)
}
}
\keyword{S3}
\keyword{classes}
