% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassMethod-XGR.r
\name{bLD}
\alias{bLD}
\alias{print.bLD}
\title{Definition for S3 class \code{bLD}}
\usage{
bLD(best, block)

\method{print}{bLD}(x, ...)
}
\arguments{
\item{best}{a GR object}

\item{block}{a GRL object}

\item{x}{an object of class \code{bLD}}

\item{...}{other parameters}
}
\value{
an object of S3 class \code{bLD}
}
\description{
\code{bLD} has 2 components: best, block.
}
\examples{
\dontrun{
# Load the library
library(XGR)
}
RData.location <- "http://galahad.well.ox.ac.uk/bigdata"
\dontrun{
bLD(best, block)
}
}
\keyword{S3}
\keyword{classes}
