% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassMethod-XGR.r
\name{cPath}
\alias{cPath}
\alias{print.cPath}
\title{Definition for S3 class \code{cPath}}
\usage{
cPath(ig_paths, gp_paths, gp_heatmap, ig_subg)

\method{print}{cPath}(x, ...)
}
\arguments{
\item{ig_paths}{an igraph object}

\item{gp_paths}{a ggplot object}

\item{gp_heatmap}{a ggplot object}

\item{ig_subg}{an igraph object}

\item{x}{an object of class \code{cPath}}

\item{...}{other parameters}
}
\value{
an object of S3 class \code{cPath}
}
\description{
\code{cPath} has 4 components: ig_paths, gp_paths, gp_heatmap, ig_subg.
}
\examples{
\dontrun{
# Load the library
library(XGR)
}
RData.location <- "http://galahad.well.ox.ac.uk/bigdata"
\dontrun{
cPath(ig_paths, gp_paths, gp_heatmap, ig_subg)
}
}
\keyword{S3}
\keyword{classes}
