% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassMethod-XGR.r
\name{DR}
\alias{DR}
\alias{print.DR}
\title{Definition for S3 class \code{DR}}
\usage{
DR(df, index, gp)

\method{print}{DR}(x, ...)
}
\arguments{
\item{df}{a data frame}

\item{index}{a data frame}

\item{gp}{a ggplot object}

\item{x}{an object of class \code{DR}}

\item{...}{other parameters}
}
\value{
an object of S3 class \code{DR}
}
\description{
\code{DR} has 3 components: df, index, gp.
}
\examples{
\dontrun{
# Load the library
library(XGR)
}
RData.location <- "http://galahad.well.ox.ac.uk/bigdata"
\dontrun{
DR(df, index, gp)
}
}
\keyword{S3}
\keyword{classes}
