% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassMethod-XGR.r
\name{EG}
\alias{EG}
\alias{print.EG}
\title{Definition for S3 class \code{EG}}
\usage{
EG(gene_info)

\method{print}{EG}(x, ...)
}
\arguments{
\item{gene_info}{a data frame}

\item{x}{an object of class \code{EG}}

\item{...}{other parameters}
}
\value{
an object of S3 class \code{EG}
}
\description{
\code{EG} has 1 component: gene_info.
}
\examples{
\dontrun{
# Load the library
library(XGR)
}
RData.location <- "http://galahad.well.ox.ac.uk/bigdata"
\dontrun{
EG(gene_info)
}
}
\keyword{S3}
\keyword{classes}
