% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassMethod-XGR.r
\name{eTerm}
\alias{eTerm}
\alias{print.eTerm}
\title{Definition for S3 class \code{eTerm}}
\usage{
eTerm(term_info, annotation, g, data, background, overlap, fc, zscore,
pvalue,
adjp, cross)

\method{print}{eTerm}(x, ...)
}
\arguments{
\item{term_info}{a data frame}

\item{annotation}{a list}

\item{g}{an 'igraph' object}

\item{data}{a vector}

\item{background}{a vector}

\item{overlap}{a vector}

\item{fc}{a vector}

\item{zscore}{a vector}

\item{pvalue}{a vector}

\item{adjp}{a vector}

\item{cross}{a matrix}

\item{x}{an object of class \code{eTerm}}

\item{...}{other parameters}
}
\value{
an object of S3 class \code{eTerm}
}
\description{
\code{eTerm} mush have following components: term_info, annotation, g,
data, background, overlap, fc, zscore, pvalue, adjp, cross.
}
\examples{
\dontrun{
# Load the library
library(XGR)
}
RData.location <- "http://galahad.well.ox.ac.uk/bigdata"
\dontrun{
eTerm(term_info, annotation, g, data, background, overlap, fc, zscore,
pvalue, adjp, cross)
}
}
\keyword{S3}
\keyword{classes}
