% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassMethod-XGR.r
\name{GS}
\alias{GS}
\alias{print.GS}
\title{Definition for S3 class \code{GS}}
\usage{
GS(set_info, gs)

\method{print}{GS}(x, ...)
}
\arguments{
\item{set_info}{a data frame}

\item{gs}{a list}

\item{x}{an object of class \code{GS}}

\item{...}{other parameters}
}
\value{
an object of S3 class \code{GS}
}
\description{
\code{GS} has 2 components: set_info, gs.
}
\examples{
\dontrun{
# Load the library
library(XGR)
}
RData.location <- "http://galahad.well.ox.ac.uk/bigdata"
\dontrun{
GS(set_info, gs)
}
}
\keyword{S3}
\keyword{classes}
