% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassMethod-XGR.r
\name{pPerf}
\alias{pPerf}
\alias{print.pPerf}
\title{Definition for S3 class \code{pPerf}}
\usage{
pPerf(data, auroc, fmax, amax, direction, gp, Pred_obj)

\method{print}{pPerf}(x, ...)
}
\arguments{
\item{data}{a data frame}

\item{auroc}{a scalar}

\item{fmax}{a scalar}

\item{amax}{a scalar}

\item{direction}{a character}

\item{gp}{a 'ggplot' object or NULL}

\item{Pred_obj}{a ROCR prediction-class object}

\item{x}{an object of class \code{pPerf}}

\item{...}{other parameters}
}
\value{
an object of S3 class \code{pPerf}
}
\description{
\code{pPerf} mush have following components: data, auroc, fmax, amax,
direction, gp, Pred_obj.
}
\examples{
\dontrun{
# Load the library
library(XGR)
}
RData.location <- "http://galahad.well.ox.ac.uk/bigdata"
\dontrun{
pPerf(data, auroc, fmax, amax, direction, gp, Pred_obj)
}
}
\keyword{S3}
\keyword{classes}
