% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassMethod-XGR.r
\name{sClass}
\alias{sClass}
\alias{print.sClass}
\title{Definition for S3 class \code{sClass}}
\usage{
sClass(prediction, predictor, performance)

\method{print}{sClass}(x, ...)
}
\arguments{
\item{prediction}{a data frame}

\item{predictor}{a data frame}

\item{performance}{a data frame}

\item{x}{an object of class \code{sClass}}

\item{...}{other parameters}
}
\value{
an object of S3 class \code{sClass}
}
\description{
\code{sClass} mush have following components: prediction, predictor,
performance.
}
\examples{
\dontrun{
# Load the library
library(XGR)
}
RData.location <- "http://galahad.well.ox.ac.uk/bigdata"
\dontrun{
sClass(prediction, predictor, performance)
}
}
\keyword{S3}
\keyword{classes}
