% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassMethod-XGR.r
\name{sMap}
\alias{sMap}
\alias{print.sMap}
\title{Definition for S3 class \code{sMap}}
\usage{
sMap(nHex, xdim, ydim, r, lattice, shape, coord, polygon, init,
codebook,
hits)

\method{print}{sMap}(x, ...)
}
\arguments{
\item{nHex}{an integer}

\item{xdim}{an integer}

\item{ydim}{an integer}

\item{r}{an integer}

\item{lattice}{a character}

\item{shape}{a character}

\item{coord}{a data frame}

\item{polygon}{a data frame}

\item{init}{a character}

\item{codebook}{a data frame}

\item{hits}{a vector}

\item{x}{an object of class \code{sMap}}

\item{...}{other parameters}
}
\value{
an object of S3 class \code{sMap}
}
\description{
\code{sClass} mush have following components: nHex, xdim, ydim, r,
lattice, shape, coord, polygon, init, codebook, hits
}
\examples{
\dontrun{
# Load the library
library(XGR)
}
RData.location <- "http://galahad.well.ox.ac.uk/bigdata"
\dontrun{
sMap(nHex=nHex, xdim=xdim, ydim=ydim, r=r, lattice=lattice,
shape=shape, coord=coord, polygon=polygon, init=init,
codebook=codebook, hits=hits)
}
}
\keyword{S3}
\keyword{classes}
