% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xAuxRd2HTML.r
\name{xAuxRd2HTML}
\alias{xAuxRd2HTML}
\title{Function to convert Rd files to HTML files}
\usage{
xAuxRd2HTML(path.from = "./XGR/man", path.to = "./XGR/vignettes")
}
\arguments{
\item{path.from}{a directory containing Rd files converted from}

\item{path.to}{a directory containing HTML files converted to}
}
\value{
none
}
\description{
\code{xAuxRd2HTML} is supposed to convert Rd files to HTML files.
}
\note{
This auxiliary function helps create a new package.
}
\examples{
# xAuxRd2HTML(path.from="./XGR/man", path.to="./XGR/vignettes")
}
\seealso{
\code{\link{xAuxRd2HTML}}
}
