% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xAuxRdWrap.r
\name{xAuxRdWrap}
\alias{xAuxRdWrap}
\title{Function to wrap texts from Rd files}
\usage{
xAuxRdWrap(path = "./XGR/man", remove.dontrun = FALSE)
}
\arguments{
\item{path}{a directory containing Rd files}

\item{remove.dontrun}{logical to indicate whether to remove the
restriction of not running examples. By default, it sets to FALSE
without any modefications}
}
\value{
none
}
\description{
\code{xAuxRdWrap} is supposed to wrap texts from Rd files under a given
directory.
}
\note{
This auxiliary function helps create a new package. The orignal Rd
files will be replaced with new ones.
}
\examples{
# xAuxRdWrap(path="./XGR/man", remove.dontrun=FALSE)
}
\seealso{
\code{\link{xAuxRdWrap}}
}
