% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xAuxSI.r
\name{xAuxSI}
\alias{xAuxSI}
\title{Function to collect information on the R session}
\usage{
xAuxSI(si = utils::sessionInfo(), file = NULL)
}
\arguments{
\item{si}{an object of the class "sessionInfo"}

\item{file}{a file name. By defaul it is NULL}
}
\value{
If the file is not NULL, it will write into the file. Otherwise it will
return a tibble with columns 'kind' ('other' other attached packages,
'loaded' for loaded but not attached packages), 'pkg' (the package
name), 'version' (the package version), 'title' (the package title) and
'description' (the package description)
}
\description{
\code{xAuxSI} is supposed to collect information on the R session. Only
attached
}
\note{
none.
}
\examples{
# res <- xAuxSI()
}
\seealso{
\code{\link{xAuxSI}}
}
