% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xBicreate.r
\name{xBicreate}
\alias{xBicreate}
\title{Function to create a bipartitle graph}
\usage{
xBicreate(data, verbose = TRUE)
}
\arguments{
\item{data}{a data frame/matrix to create a bipartitle graph. The
coloring can be continuous (numeric matrix) or discrete (factor
matrix)}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to true for display}
}
\value{
It returns an igraph object.
}
\description{
\code{xBicreate} is supposed to create a bipartitle graph.
}
\note{
none
}
\examples{
data <- cbind(c(1,1,0,1), c(0,1,2,0))
ig <- xBicreate(data)
}
\seealso{
\code{\link{xBicreate}}
}
