% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCEMiTool.r
\name{xCEMiTool}
\alias{xCEMiTool}
\title{Function to post-process results from CEMiTool}
\usage{
xCEMiTool(cem, filename = "xCEMiTool", verbose = T)
}
\arguments{
\item{cem}{an object of class "CEMiTool"}

\item{filename}{the without-extension part of the name of the output
file. By default, it is 'xCEMiTool'}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to true for display}
}
\value{
a data frame with three columns 'genes' (ordered by intramodular
connectivity), 'modules' (M1, M2, ...) and 'rank' (integer with a
module).
}
\description{
\code{xCEMiTool} is supposed to post-process results from CEMiTool (an
object of class "CEMiTool").
}
\note{
none
}
\examples{
\dontrun{
# Load the library
library(XGR)
RData.location <- "http://galahad.well.ox.ac.uk/bigdata/"

cem <- cemitool(expr, filter=FALSE, merge_similar=TRUE, min_ngen=70,
diss_thresh=0.8, verbose=T)
df_res <- xCEMiTool(cem)
}
}
\seealso{
\code{\link{xCEMiTool}}
}
