% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xClassifyCV.r
\name{xClassifyCV}
\alias{xClassifyCV}
\title{Function to visualise cross-validation performance}
\usage{
xClassifyCV(list_sClass, displayBy = c("ROC", "Fmax"), type = c("bar",
"ridge"), sort = TRUE, colormap = "spectral", ncolors = 64, ...)
}
\arguments{
\item{list_sClass}{a list of "sClass" objects or an object of class
"sClass"}

\item{displayBy}{which statistics will be used for displaying. It can
be "ROC" for AUC in ROC (shaped by direction), "Fmax" for F-max in
Precision-Recall curve}

\item{type}{the plot type. It can be "bar" and "ridge"}

\item{sort}{logical to indicate whether to sort methods according to
performance. By default, it sets TRUE}

\item{colormap}{short name for the colormap. It can be one of "jet"
(jet colormap), "bwr" (blue-white-red colormap), "gbr" (green-black-red
colormap), "wyr" (white-yellow-red colormap), "br" (black-red
colormap), "yr" (yellow-red colormap), "wb" (white-black colormap), and
"rainbow" (rainbow colormap, that is,
red-yellow-green-cyan-blue-magenta). Alternatively, any
hyphen-separated HTML color names, e.g. "blue-black-yellow",
"royalblue-white-sandybrown", "darkgreen-white-darkviolet". A list of
standard color names can be found in
\url{http://html-color-codes.info/color-names}}

\item{ncolors}{the number of colors specified over the colormap}

\item{...}{additional graphic parameters for
ggridges::geom_density_ridges_gradient (such as scale=1)}
}
\value{
an object of class "ggplot"
}
\description{
\code{xClassifyCV} is supposed to visualise cross-validation
performance. It returns an object of class "ggplot".
}
\note{
none
}
\examples{
\dontrun{
# Load the library
library(XGR)
}
RData.location <- "http://galahad.well.ox.ac.uk/bigdata"
\dontrun{
gp <- xClassifyCV(list_sClass, displayBy="ROC", type="bar")
gp <- xClassifyCV(list_sClass, displayBy="ROC", type="ridge")
gp <- xClassifyCV(sClass)
}
}
\seealso{
\code{\link{xClassifyCV}}
}
