% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xClassifyPara.r
\name{xClassifyPara}
\alias{xClassifyPara}
\title{Function to visualise tuning parameters}
\usage{
xClassifyPara(sClass, colormap = "spectral", ncolors = 50,
contour = TRUE, plot.3D = FALSE, theta.3D = 40, phi.3D = 25,
text.3D = TRUE)
}
\arguments{
\item{sClass}{an object of class "sClass". Alternatively, it can be a
data frame}

\item{colormap}{short name for the colormap. It can be one of "jet"
(jet colormap), "bwr" (blue-white-red colormap), "gbr" (green-black-red
colormap), "wyr" (white-yellow-red colormap), "br" (black-red
colormap), "yr" (yellow-red colormap), "wb" (white-black colormap), and
"rainbow" (rainbow colormap, that is,
red-yellow-green-cyan-blue-magenta). Alternatively, any
hyphen-separated HTML color names, e.g. "blue-black-yellow",
"royalblue-white-sandybrown", "darkgreen-white-darkviolet". A list of
standard color names can be found in
\url{http://html-color-codes.info/color-names}}

\item{ncolors}{the number of colors specified over the colormap}

\item{contour}{logical to indicate whether 2D coutour plot should be
also included}

\item{plot.3D}{logical to indicate whether to plot in 3D}

\item{theta.3D}{the azimuthal direction. By default, it is 40}

\item{phi.3D}{the colatitude direction. By default, it is 20}

\item{text.3D}{logical whether to text the point with the maximum
value}
}
\value{
a ggplot object (1D), a trellis object (2D), and NULl (3D)
}
\description{
\code{xClassifyPara} is supposed to visualise tuning parameters.
}
\note{
none
}
\examples{
\dontrun{
library(XGR)
}
RData.location <- "http://galahad.well.ox.ac.uk/bigdata"
\dontrun{
gp <- xClassifyPara(sClass)
xClassifyPara(sClass, plot.3D=T, theta.3D=40, phi.3D=45)
}
}
\seealso{
\code{\link{xClassifyPara}}
}
