% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xClassifyPred.r
\name{xClassifyPred}
\alias{xClassifyPred}
\title{Function to visualise predictors}
\usage{
xClassifyPred(sClass, displayBy = c("ROC", "Fmax", "importance",
"Amax"),
sort = TRUE)
}
\arguments{
\item{sClass}{an object of class "sClass"}

\item{displayBy}{which statistics will be used for displaying. It can
be "importance" for predictor importance, "ROC" for AUC in ROC (shaped
by direction), "Fmax" for F-max in Precision-Recall curve, and "Amax"
for maximum accuracy}

\item{sort}{logical to indicate whether to sort methods according to
performance. By default, it sets TRUE}
}
\value{
an object of class "ggplot"
}
\description{
\code{xClassifyPred} is supposed to visualise predictors from a
'sClass' object. It returns an object of class "ggplot".
}
\note{
none
}
\examples{
\dontrun{
# Load the library
library(XGR)
}
RData.location <- "http://galahad.well.ox.ac.uk/bigdata"
\dontrun{
gp <- xClassifyPred(sClass, displayBy="ROC")
gp <- xClassifyPred(sClass, displayBy="importance")
}
}
\seealso{
\code{\link{xClassifyPred}}
}
