% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCpartition.r
\name{xCpartition}
\alias{xCpartition}
\title{Function to draw a partition-like circular plot}
\usage{
xCpartition(ig, leave.label.direction = c("none", "leftright",
"topbottom"), leave.label.size = 2, leave.label.color = "darkblue",
leave.label.alpha = 0.7, leave.label.wrap = NULL,
leave.label.offset = 0.5, leave.size = 0, limit.expansion = NULL,
edge.color = "grey", edge.alpha = 0.5, edge.width = 0.5, ...)
}
\arguments{
\item{ig}{an object of class "igraph" with node attribute 'name'. It
could be a 'phylo' object converted to. Note: the leave labels would be
the node attribute 'name' unless the node attribute 'label' is
explicitely provided}

\item{leave.label.direction}{the leave label direction. It can be
"none", "leftright" (aligned to the left- and right-most edge) and
"topbottom" (aligned to the top- and bottom-most edge)}

\item{leave.label.size}{the text size of the leave labelings. By
default, it is 2}

\item{leave.label.color}{the color of the leave labelings}

\item{leave.label.alpha}{the alpha of the leave labelings}

\item{leave.label.wrap}{the wrap width of the leave labelings}

\item{leave.label.offset}{the offset of the leave labelings aligned to
the edge. It is defined as relative to the range of limits (x-limit for
left-right, and y-limit for top-bottom)}

\item{leave.size}{the size of the leave nodes. By default, it is 0}

\item{limit.expansion}{the x- and y-limit expansion. By default, it is
NULL, decided by "leave.label.offset"}

\item{edge.color}{the color of edges}

\item{edge.alpha}{the alpha of edges}

\item{edge.width}{the width of edges}

\item{...}{additional graphic parameters (such as size, color) used in
ggrepel::geom_text_repel to control labels}
}
\value{
a ggplot2 object appended with 'ig' and 'data' which should contain
columns 'x','y', 'leaf' (T/F), 'depth' (the number of step to the
root), 'name' (the same as V(ig)$name), 'tipid' (tip id), 'label' (if
not given in ig, a 'name' varient).
}
\description{
\code{xCpartition} is supposed to draw a partition-like circular plot
(partition circular layout), with tips labelled (aligned to left-right
or top-bottom edges).
}
\note{
none
}
\examples{
\dontrun{
library(XGR)
RData.location <- "http://galahad.well.ox.ac.uk/bigdata"

AA.path <- xRDataLoader("AA.path", RData.location=RData.location)
# consensus tree
ig <- AA.path$consensus$ig

# none
gp <- xCpartition(ig, leave.label.direction="none",
leave.label.wrap=50, leave.label.offset=0.5, leave.size=2)
head(gp$data \%>\% dplyr::arrange(tipid))

# leftright
gp <- xCpartition(ig, leave.label.direction="leftright",
leave.label.wrap=50, leave.label.offset=0.5, leave.size=2)

# topbottom
gp <- xCpartition(ig, leave.label.direction="topbottom",
leave.label.wrap=50, leave.label.offset=0.5, leave.size=2)
}
}
\seealso{
\code{\link{xCpartition}}
}
