% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xEnrichConciser.r
\name{xEnrichConciser}
\alias{xEnrichConciser}
\title{Function to make enrichment results conciser by removing redundant terms}
\usage{
xEnrichConciser(eTerm, cutoff = 0.95)
}
\arguments{
\item{eTerm}{an object of class "eTerm"}

\item{cutoff}{the cutoff used to remove redunant terms. For a term
(less significant), if there is a more significant term having members
that cover at least the cutoff percentage (0.95 by default) of its
members, then this term will be removed}
}
\value{
an object of class "eTerm", after redundant terms being removed.
}
\description{
\code{xEnrichConciser} is supposed to make enrichment results conciser
by removing redundant terms.
}
\note{
none
}
\examples{
\dontrun{
eTerm_concise <- xEnrichConciser(eTerm)
}
}
\seealso{
\code{\link{xEnricherGenes}}, \code{\link{xEnricherSNPs}}
}

