% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xEnricherRegions.r
\name{xEnricherRegions}
\alias{xEnricherRegions}
\title{Function to conduct region-based enrichment analysis using genomic annotations}
\usage{
xEnricherRegions(data.file, annotation.file = NULL, background.file =
NULL,
format.file = c("data.frame", "bed", "chr:start-end"),
background.annotatable.only = T, test = c("hypergeo", "fisher",
"binomial"), p.adjust.method = c("BH", "BY", "bonferroni", "holm",
"hochberg", "hommel"), GR.annotation = c(NA, "Uniform_TFBS",
"Uniform_DNaseI_HS", "Broad_Histone", "SYDH_Histone", "UW_Histone",
"FANTOM5_Enhancer_Cell", "FANTOM5_Enhancer_Tissue",
"FANTOM5_Enhancer_Extensive", "FANTOM5_Enhancer",
"Segment_Combined_Gm12878",
"Segment_Combined_H1hesc", "Segment_Combined_Helas3",
"Segment_Combined_Hepg2", "Segment_Combined_Huvec",
"Segment_Combined_K562",
"TFBS_Conserved", "TS_miRNA", "TCGA"), verbose = T,
RData.location =
"https://github.com/hfang-bristol/RDataCentre/blob/master/Portal")
}
\arguments{
\item{data.file}{an input data file, containing a list of genomic
regions to test. If the input file is formatted as a 'data.frame'
(specified by the parameter 'format.file' below), the first three
columns correspond to the chromosome (1st column), the starting
chromosome position (2nd column), and the ending chromosome position
(3rd column). If the format is indicated as 'bed' (browser extensible
data), the same as 'data.frame' format but the position is 0-based
offset from chromomose position. If the genomic regions provided are
not ranged but only the single position, the ending chromosome position
(3rd column) is allowed not to be provided. If the format is indicated
as "chr:start-end", instead of using the first 3 columns, only the
first column will be used and processed. If the file also contains
other columns, these additional columns will be ignored. Alternatively,
the input file can be the content itself assuming that input file has
been read. Note: the file should use the tab delimiter as the field
separator between columns}

\item{annotation.file}{an input annotation file containing genomic
annotations for genomic regions. If the input file is formatted as a
'data.frame', the first four columns correspond to the chromosome (1st
column), the starting chromosome position (2nd column), the ending
chromosome position (3rd column), and the genomic annotations (eg
transcription factors and histones; 4th column). If the format is
indicated as 'bed', the same as 'data.frame' format but the position is
0-based offset from chromomose position. If the format is indicated as
"chr:start-end", the first two columns correspond to the
chromosome:start-end (1st column) and the genomic annotations (eg
transcription factors and histones; 2nd column). If the file also
contains other columns, these additional columns will be ignored.
Alternatively, the input file can be the content itself assuming that
input file has been read. Note: the file should use the tab delimiter
as the field separator between columns}

\item{background.file}{an input background file containing a list of
genomic regions as the test background. The file format is the same as
'data.file'. By default, it is NULL meaning all annotatable bases (ig
non-redundant bases covered by 'annotation.file') are used as
background. However, if only one annotation (eg only a transcription
factor) is provided in 'annotation.file', the background must be
provided.}

\item{format.file}{the format for input files. It can be one of
"data.frame", "chr:start-end", "bed"}

\item{background.annotatable.only}{logical to indicate whether the
background is further restricted to annotatable bases (covered by
'annotation.file'). In other words, if the background is provided, the
background bases are those after being overlapped with annotatable
bases. Notably, if only one annotation (eg only a transcription factor)
is provided in 'annotation.file', it should be false.}

\item{test}{the statistic test used. It can be "fisher" for using
fisher's exact test, "hypergeo" for using hypergeometric test, or
"binomial" for using binomial test. Fisher's exact test is to test the
independence between gene group (genes belonging to a group or not) and
gene annotation (genes annotated by a term or not), and thus compare
sampling to the left part of background (after sampling without
replacement). Hypergeometric test is to sample at random (without
replacement) from the background containing annotated and non-annotated
genes, and thus compare sampling to background. Unlike hypergeometric
test, binomial test is to sample at random (with replacement) from the
background with the constant probability. In terms of the ease of
finding the significance, they are in order: hypergeometric test >
binomial test > fisher's exact test. In other words, in terms of the
calculated p-value, hypergeometric test < binomial test < fisher's
exact test}

\item{p.adjust.method}{the method used to adjust p-values. It can be
one of "BH", "BY", "bonferroni", "holm", "hochberg" and "hommel". The
first two methods "BH" (widely used) and "BY" control the false
discovery rate (FDR: the expected proportion of false discoveries
amongst the rejected hypotheses); the last four methods "bonferroni",
"holm", "hochberg" and "hommel" are designed to give strong control of
the family-wise error rate (FWER). Notes: FDR is a less stringent
condition than FWER}

\item{GR.annotation}{the genomic regions of annotation data. By
default, it is 'NA' to disable this option. Predefined annotation data
include: ENCODE Transcription Factor ChIP-seq Uniform Peaks data
("Uniform_TFBS"), ENCODE DNaseI Hypersensitivity Uniform Peaks data
("Uniform_DNaseI_HS"), ENCODE Histone Modification ChIP-seq data from
different sources ("Broad_Histone","SYDH_Histone","UW_Histone"),
FANTOM5 expressed enhancer altas
("FANTOM5_Enhancer_Cell","FANTOM5_Enhancer_Tissue","FANTOM5_Enhancer_Extensive","FANTOM5_Enhancer"),
ENCODE combined Genome Segmentation data (ChromHMM and Segway) in 6
cell lines
("Segment_Combined_Gm12878","Segment_Combined_H1hesc","Segment_Combined_Helas3","Segment_Combined_Hepg2","Segment_Combined_Huvec","Segment_Combined_K562"),
the human/mouse/rat conserved TFBS ("TFBS_Conserved"), TargetScan miRNA
regulatory sites ("TS_miRNA"), TCGA exome mutation data across 12 tumor
types ("TCGA"). Beyond these predefined data, the user can specify the
customised input. To do so, first save your RData file (a list of GR
objects, each is an GR object correponding to an annotation) into your
local computer. Then, tell "GR.annotation" with your RData file name
(with or without extension), plus specify your file RData path in
"RData.location"}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to false for no display}

\item{RData.location}{the characters to tell the location of built-in
RData files. See \code{\link{xRDataLoader}} for details}
}
\value{
a data frame with 8 columns:
\itemize{
\item{\code{name}: the annotation name}
\item{\code{nAnno}: the number of nucleotides/bases covered by that
annotation. If the background is provided, they are also restricted by
this}
\item{\code{nOverlap}: the number of nucleotides/bases overlapped
between input regions and annotation regions. If the background is
provided, they are also restricted by this}
\item{\code{zscore}: z-score}
\item{\code{pvalue}: p-value}
\item{\code{adjp}: adjusted p-value. It is the p value but after being
adjusted for multiple comparisons}
\item{\code{expProb}: the probability of expecting nucleotides/bases
overlapped between background regions and annotation regions}
\item{\code{obsProb}: the probability of observing nucleotides/bases
overlapped between input regions and annotation regions}
}
}
\description{
\code{xEnricherRegions} is supposed to conduct region-based enrichment
analysis for the input genomic region dataan, using genomic annotations
(eg active chromatin, transcription factor binding sites/motifs,
conserved sites). Enrichment analysis is based on either Fisher's exact
test or Hypergeometric test for estimating the significance of overlaps
at the single base-pair resolution. Test background can be provided; by
default, the annotatable bases will be used.
}
\note{
None
}
\examples{
\dontrun{
# Load the library
library(XGR)

# Enrichment analysis for GWAS SNPs from ImmunoBase
# a) provide input data
data.file <- "http://galahad.well.ox.ac.uk/bigdata/ImmunoBase_GWAS.bed"

# b) perform enrichment analysis using FANTOM expressed enhancers
eTerm <- xEnricherRegions(data.file=data.file, format.file="bed",
GR.annotation="FANTOM5_Enhancer_Cell")

# c) view enrichment results for the top significant terms
xEnrichViewer(eTerm)

# d) save enrichment results to the file called 'Regions_enrichments.txt'
output <- xEnrichViewer(eTerm, top_num=length(eTerm$adjp),
sortBy="adjp", details=TRUE)
utils::write.table(output, file="Regions_enrichments.txt", sep="\\t",
row.names=FALSE)
}
}
\seealso{
\code{\link{xEnrichViewer}}
}

