% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xGGtree.r
\name{xGGtree}
\alias{xGGtree}
\title{Function to visualise enrichment results using a forest plot}
\usage{
xGGtree(phylo, membership, layout = c("rectangular", "fan"),
colormap = "ggplot2", extendto = NULL, offset = NULL,
open.angle = 0)
}
\arguments{
\item{phylo}{an object of class "phylo"}

\item{membership}{a named input vector containing the membership for
tips. For this named vector, the element names are clusters, the
element values for tip labels}

\item{layout}{the visual layout. It can be "rectangular" or "fan"}

\item{colormap}{short name for the colormap. It can be one of "jet"
(jet colormap), "bwr" (blue-white-red colormap), "gbr" (green-black-red
colormap), "wyr" (white-yellow-red colormap), "br" (black-red
colormap), "yr" (yellow-red colormap), "wb" (white-black colormap), and
"rainbow" (rainbow colormap, that is,
red-yellow-green-cyan-blue-magenta). Alternatively, any
hyphen-separated HTML color names, e.g. "blue-black-yellow",
"royalblue-white-sandybrown", "darkgreen-white-darkviolet". A list of
standard color names can be found in
\url{http://html-color-codes.info/color-names}}

\item{extendto}{the maximum value of x. By default it is NULL. Only
works for the rectangular visual}

\item{offset}{the offset to the tips. By default it is NULL. Only works
for the rectangular visual}

\item{open.angle}{the angle for the fan visual}
}
\value{
an object of class "ggplot" appended with a phylo object "tree"
(appended with a "cluster" list)
}
\description{
\code{xGGtree} is supposed to visualise enrichment results using a
forest plot. A point is colored by the significance level, and a
horizontal line for the 95% confidence interval (CI) of odds ratio (OR;
the wider the CI, the less reliable). It returns an object of class
"ggplot".
}
\note{
none
}
\examples{
\dontrun{
# Load the library
library(XGR)
gp <- xGGtree(tree, membership)
}
}
\seealso{
\code{\link{xGGtree}}
}
