\name{XGR-package}
\alias{XGR-package}
\alias{XGR}
\docType{package}
\title{
\packageTitle{XGR}
}
\description{
\packageDescription{XGR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{XGR}
\packageIndices{XGR}
~~ An overview of how to use the package, including the most important
~~
~~ functions ~~
}
\author{
\packageAuthor{XGR}

Maintainer: \packageMaintainer{XGR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS
in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
