% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xGraphML.r
\name{xGraphML}
\alias{xGraphML}
\title{Function to visualise a graph object of class "igraph"}
\usage{
xGraphML(g, pattern = NULL, colormap = "wyr", ncolors = 64, zlim =
NULL,
filename = "xGraphML")
}
\arguments{
\item{g}{an object of class "igraph"}

\item{pattern}{a numeric vector used to color-code vertices/nodes.
Notably, if the input vector contains names, then these names should
include all node names of input graph, i.e. V(g)$name, since there is a
mapping operation. After mapping, the length of the patern vector
should be the same as the number of nodes of input graph; otherwise,
this input pattern will be ignored. The way of how to color-code is to
map values in the pattern onto the whole colormap (see the next
arguments: colormap, ncolors, zlim and colorbar)}

\item{colormap}{short name for the colormap. It can be one of "jet"
(jet colormap), "bwr" (blue-white-red colormap), "gbr" (green-black-red
colormap), "wyr" (white-yellow-red colormap), "br" (black-red
colormap), "yr" (yellow-red colormap), "wb" (white-black colormap),
"rainbow" (rainbow colormap, that is,
red-yellow-green-cyan-blue-magenta), and "ggplot2" (emulating ggplot2
default color palette). Alternatively, any hyphen-separated HTML color
names, e.g. "lightyellow-orange" (by default), "blue-black-yellow",
"royalblue-white-sandybrown", "darkgreen-white-darkviolet". A list of
standard color names can be found in
\url{http://html-color-codes.info/color-names}}

\item{ncolors}{the number of colors specified over the colormap}

\item{zlim}{the minimum and maximum z/patttern values for which colors
should be plotted, defaulting to the range of the finite values of z.
Each of the given colors will be used to color an equispaced interval
of this range. The midpoints of the intervals cover the range, so that
values just outside the range will be plotted}

\item{filename}{the without-extension part of the name of the output
file. By default, it is 'xGraphML'}
}
\value{
invisible
}
\description{
\code{xGraphML} is supposed to visualise a graph object of class
"igraph". It also allows vertices/nodes color-coded according to the
input pattern.
}
\note{
none
}
\examples{
\dontrun{
# Load the library
library(XGR)
RData.location <- "http://galahad.well.ox.ac.uk/bigdata_dev/"

# load REACTOME
# restricted to Immune System ('R-HSA-168256') or Signal Transduction ('R-HSA-162582')
g <- xRDataLoader(RData.customised='ig.REACTOME',
RData.location=RData.location)
neighs.out <- igraph::neighborhood(g, order=vcount(g),
nodes="R-HSA-168256", mode="out")
nodeInduced <- V(g)[unique(unlist(neighs.out))]$name
ig <- igraph::induced.subgraph(g, vids=nodeInduced)

# 4) visualise the graph with vertices being color-coded by the pattern
pattern <- runif(vcount(ig))
names(pattern) <- V(ig)$name
xGraphML(g=ig, pattern=pattern, colormap="bwr")
}
}
\seealso{
\code{\link{xGraphML}}
}
