% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xGraphPerf.r
\name{xGraphPerf}
\alias{xGraphPerf}
\title{Function to evaluate the graph node similarity prediction performance}
\usage{
xGraphPerf(prediction, g = NULL, GSP = NULL, GSN = NULL,
plot = c("none", "ROC", "PR"), highlight = F, verbose = TRUE)
}
\arguments{
\item{prediction}{a symmetric sparse matrix containing predictive
scores}

\item{g}{an object of class "igraph" (or "graphNEL") for a graph. If
either GSP or GSN is NULL, its adjacency matrix is used to define GSP
(all edges) and the rest as GSN}

\item{GSP}{a vector containing Gold Standard Positives (GSP)}

\item{GSN}{a vector containing Gold Standard Negatives (GSN)}

\item{plot}{the way to plot performance curve. It can be 'none' for no
curve returned, 'ROC' for ROC curve, and 'PR' for PR curve.}

\item{highlight}{logical to indicate whether a dot is highlighted. It
only works when plot is drawn. When true, the maximum accuracy
highlighted in ROC curve, and the Fmax highlighted in PR curve. By
default, it sets to false}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to TRUE for display}
}
\value{
an object of class "pPerf", a list with following components:
\itemize{
\item{\code{data}: a data frame with 8 columns, including 4 performance
measures ('Accuracy', 'Precision', 'Recall' and 'Specificity'), 'name'
(subjects), 'pred' (predictive scores), 'label' (1 for GSP and 0 for
GSN), 'corrected' (corrected/transformed predictiv scores, always the
higher the better)}
\item{\code{auroc}: a scalar value for ROC AUC}
\item{\code{fmax}: a scalar value for maximum F-measure}
\item{\code{amax}: a scalar value for maximum accuracy}
\item{\code{direction}: '+' (the higher score the better prediction)
and '-' (the higher score the worse prediction)}
\item{\code{gp}: a ggplot object (if plotted) or NULL}
\item{\code{Pred_obj}: a ROCR prediction-class object (potentially used
for calculating other performance measures)}
}
}
\description{
\code{xGraphPerf} is supposed to evaluate the graph node similarity
prediction performance. It returns an object of class "pPerf".
}
\note{
AUC: the area under ROC
F-measure: the maximum of a harmonic mean between precision and recall
along PR curve
}
\examples{
\dontrun{
# Load the library
library(XGR)
}
RData.location <- "http://galahad.well.ox.ac.uk/bigdata"
\dontrun{
g <- make_graph("Zachary")
# Common Neighbors (CN)
prediction <- xGraphSim(g, measure="CN")
pPerf <- xGraphPerf(prediction, g)
gp <- xClassifyComp(pPerf, displayBy="PR")
}
}
