% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xGraphSim.r
\name{xGraphSim}
\alias{xGraphSim}
\title{Function to calculate node similarity based on the network structure}
\usage{
xGraphSim(g, measure = c("CN", "SI", "JI", "DI", "HPI", "HDI", "LHN",
"AA", "RA", "LP", "KI", "LHNg", "SP", "ACT", "CL", "MFI", "RWR",
"SRW"),
type = c("full", "edge"), edge.fast = F, measure.para = list(LP.eps
= 0.01, KI.beta = 0.001, LHNg.theta = 0.5, RWR.r = 0.5, SRW.p = 4),
verbose = TRUE)
}
\arguments{
\item{g}{an object of class "igraph" (or "graphNEL") for a graph}

\item{measure}{the similarity metrics used to measure similarity
between nodes. It can be node-dependent similarity, path-dependent
similarity, and random walk-based similarity}

\item{type}{the type defining nodes to be calcuated. It can be 'full'
for all possible pairs of nodes (used for link prediction) and 'edge'
for paris of nodes in edges (used for link recommendation)}

\item{edge.fast}{logical to indicate whether a vectorised fast
computation is used. By default, it sets to true. It is always
advisable to use this vectorised fast computation; though the
conventional computation is just used for script understanding}

\item{measure.para}{a list of measure-specific parameters.}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to true for display}
}
\value{
It returns a symmetric sparse matrix involving all network nodes.
}
\description{
\code{xGraphSim} is supposed to calculate node similarity based on the
network structure.
}
\note{
The similarity measures are detailed as follows.\cr
1. Node-dependent similarity
\itemize{
\item{"Common Neighbors (CN)": the number of common neighbors.}
\item{"Salton Index (SI)": the cosine similarity, that is, the number
of common neighbors normalised by square root of degree product of two
nodes.}
\item{"Jaccard Index (JI)": the number of common neighbors normalised
by the union of neighbors.}
\item{"Dice Index (DI)": mainly for ecological community data, that is,
the number of common neighbors normalised by degree average of two
nodes.}
\item{"Hub Promoted Index (HPI)": the topological overlap, that is, the
number of common neighbors normalised by minimum degree of two nodes
(with hubs receiving higher score).}
\item{"Hub Depressed Index (HPI)": the number of common neighbors
normalised by maximum degree of two nodes (with hubs receiving lower
score).}
\item{"Leicht-Holme-Newman Index (LHN)": inversely proportional to the
expected number of common neighbors in the configuration model, that
is, the number of common neighbors normalised by degree product of two
nodes.}
\item{"Adamic-Adar Index (AA)": the number of common neighbors with a
logarithmic weighting, that is, weighted by 1/log(degree).}
\item{"Resource Allocation Index (RA)": the number of common neighbors
with an inverse weighting, that is, weighted by 1/degree.}
}
2. Path-dependent similarity
\itemize{
\item{"Local Path Index (LP)": the number of two-step paths and
three-step paths (weighted by the eps).}
\item{"Katz Index (KI)": sums over the collection of paths
exponentially damped by length to give the shorter paths more weights.}
\item{"Leicht-Holme-Newman Index global similarity (LHNg)": two nodes
are similar if their immediate neighbors are similar by themselves.}
\item{"Shortest Path Index (SP)": the length of the shortest paths.}
}
3. Random walk-based similarity
\itemize{
\item{"Average Commute Time (ACT)": the average number of steps
required by a random walker starting from a node to reach another
node.}
\item{"Cosine based on Laplacian matrix (CL)": the cosine of the angle
between columns of the pseudoinverse of the Laplacian matrix.}
\item{"Matrix Forest Index (MFI)": the ratio of the number of spanning
rooted forests belonging to the same tree to all spanning rooted
forests of the network.}
\item{"Random Walk with Restart (RWR)": the probability at the steady
state, solved analytically (not iteratively).}
\item{"Short Random Walk (SRW)": p-step random walk, solved
analytically (not iteratively).}
}
}
\examples{
g <- make_graph("Zachary")
gp <- xGGnetwork(g,node.label='name',node.label.size=3,
node.label.force=0.01)

\dontrun{
# Common Neighbors (CN)
res <- xGraphSim(g, measure="CN")
# Adamic-Adar Index (AA)
res <- xGraphSim(g, measure="AA")
# Local Path Index (LP)
res <- xGraphSim(g, measure="LP", measure.para=list(LP.eps=0.01))
# Katz Index (KI)
res <- xGraphSim(g, measure="KI", measure.para=list(KI.beta=0.001))
# Random Walk with Restart (RWR)
res <- xGraphSim(g, measure="RWR", measure.para=list(RWR.r=0.5))
# Short Random Walk (SRW)
res <- xGraphSim(g, measure="SRW", measure.para=list(SRW.p=4))

# comparing performance
measures <- c("CN","SI","JI","DI","HPI","HDI","LHN","AA","RA")
names(measures) <- measures
ls_pPerf <- lapply(measures, function(measure){
prediction <- xGraphSim(g, measure)
pPerf <- xGraphPerf(prediction, g)
})
gp <- xClassifyComp(ls_pPerf, displayBy="PR")

###################
# advanced use
###################
ig.KEGG.list <- xRDataLoader('ig.KEGG.list',
RData.location=RData.location)

## prediction within a bigraph (ligend-receptor interaction)
# 1) define bigraph (ligend-receptor interaction)
ig <- ig.KEGG.list$'Cytokine-cytokine receptor interaction'
df_edges <- igraph::get.data.frame(ig, what="edges")[, c(1,2)]
big <- xBicreate(xSparseMatrix(df_edges))
V(big)$type <- ifelse(V(big)$type=='ynode', 'Ligand', 'Receptor')
# 2) prediction 
# via Short Random Walk (SRW)
prediction <- xGraphSim(big, measure="SRW", measure.para=list(SRW.p=4))
visHeatmapAdv(as.matrix(prediction),dendrogram="none",
colormap='white-orange-red', labRow=NA, labCol=NA)
# 3) performance against GSP (the observed), GSN (any not observed)
pPerf <- xGraphPerf(prediction, big)

## prediction based on KEGG pathways
# 1) define merged graph
networks <- c("KEGG_environmental","KEGG_organismal")
ls_ig <- lapply(networks, xDefineNet, RData.location=RData.location)
mg <- xCombineNet(ls_ig, combineBy='union', attrBy="intersect")
# 2) prediction
# via Short Random Walk (SRW)
prediction <- xGraphSim(mg, measure="SRW", measure.para=list(SRW.p=4))
# 3) define the target space (ligend-receptor interaction) to evaluate
ig <- ig.KEGG.list$'Cytokine-cytokine receptor interaction'
df_edges <- igraph::get.data.frame(ig, what="edges")
# GSP (LR interactions observed)
GSP <- paste0(df_edges[,1],'--',df_edges[,2])
# GSN (any LR interaction not observed)
GSN <- apply(expand.grid(unique(df_edges[,1]), unique(df_edges[,2])),
1, paste0, collapse="--")
GSN <- base::setdiff(GSN, GSP)
# 4) performance
pPerf <- xGraphPerf(prediction, GSP=GSP, GSN=GSN)
}
}
\seealso{
\code{\link{xGraphSim}}
}
