% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xGT.r
\name{xGT}
\alias{xGT}
\title{Function to plot a tree and its clades}
\usage{
xGT(phylo, membership = NULL, layout = c("rectangular", "fan"),
tip.label = T, tip.size = 2, tip.align = T, colormap = "ggplot2",
extendto = NULL, offset = NULL, open.angle = 0)
}
\arguments{
\item{phylo}{an object of class "phylo"}

\item{membership}{a named input vector containing the membership for
tips. For this named vector, the element names are clusters, the
element values for tip labels. It can be NULL}

\item{layout}{the visual layout. It can be "rectangular" or "fan"}

\item{tip.label}{logical to indicate whether the tree tip is labelled}

\item{tip.size}{the size of the tree tip label}

\item{tip.align}{logical to indicate whether the tree tip label is
aligned}

\item{colormap}{short name for the colormap. It can be one of "jet"
(jet colormap), "bwr" (blue-white-red colormap), "gbr" (green-black-red
colormap), "wyr" (white-yellow-red colormap), "br" (black-red
colormap), "yr" (yellow-red colormap), "wb" (white-black colormap), and
"rainbow" (rainbow colormap, that is,
red-yellow-green-cyan-blue-magenta). Alternatively, any
hyphen-separated HTML color names, e.g. "blue-black-yellow",
"royalblue-white-sandybrown", "darkgreen-white-darkviolet". A list of
standard color names can be found in
\url{http://html-color-codes.info/color-names}}

\item{extendto}{the maximum value of x. By default it is NULL. Only
works for the rectangular visual}

\item{offset}{the offset to the tips. By default it is NULL. Only works
for the rectangular visual}

\item{open.angle}{the angle for the fan visual}
}
\value{
an object of class "ggplot" appended with a phylo object "tree"
(appended with a list "cluster")
}
\description{
\code{xGT} is supposed to plot a tree and its clades. It returns an
object of class "ggplot".
}
\note{
none
}
\examples{
\dontrun{
set.seed(825)
tree <- ape::rtree(50)
gp <- xGT(tree)
gp + ggtree::geom_tiplab(size=2,align=T)
gp$data
gp$tree

# fan layout
gp <- xGT(tree, layout='fan')
gp
}
}
\seealso{
\code{\link{xGT}}
}
