% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xGTheatmapAdv.r
\name{xGTheatmapAdv}
\alias{xGTheatmapAdv}
\title{Function to plot tree and heatmap given associated data on tree tips}
\usage{
xGTheatmapAdv(gp, data, ratio.width = 1, gap.width = NULL,
colormap = "spectral", ncolors = 64, zlim = NULL, barwidth = 0.5,
barheight = 4, nbin = 64, legend.title = "",
legend.text.size = 5, legend.title.size = 6, y.text.size = 2,
x.rotate = 0, x.text.size = 2, x.text.hjust = 0.5, ...)
}
\arguments{
\item{gp}{a ggplot object resulting from ggtree}

\item{data}{a data frame/matrix for coloring}

\item{ratio.width}{the width of the heatmap relative to the tree. Be
default this ratio is 1 (that is, the equal width as the tree)}

\item{gap.width}{the gap between the heatmap and the tree. Be default
it is 0.1 multiplied tree width}

\item{colormap}{short name for the colormap. It can be one of "jet"
(jet colormap), "bwr" (blue-white-red colormap), "gbr" (green-black-red
colormap), "wyr" (white-yellow-red colormap), "br" (black-red
colormap), "yr" (yellow-red colormap), "wb" (white-black colormap), and
"rainbow" (rainbow colormap, that is,
red-yellow-green-cyan-blue-magenta). Alternatively, any
hyphen-separated HTML color names, e.g. "blue-black-yellow",
"royalblue-white-sandybrown", "darkgreen-white-darkviolet". A list of
standard color names can be found in
\url{http://html-color-codes.info/color-names}}

\item{ncolors}{the number of colors specified over the colormap}

\item{zlim}{the minimum and maximum z values for which colors should be
plotted, defaulting to the range of the finite values of displayed
matrix}

\item{barwidth}{the width of the colorbar. Default value is
'legend.key.width' or 'legend.key.size' in 'theme' or theme}

\item{barheight}{the height of the colorbar. Default value is
'legend.key.height' or 'legend.key.size' in 'theme' or theme}

\item{nbin}{the number of bins for drawing colorbar}

\item{legend.title}{the title of the colorbar. By default, it is ''}

\item{legend.text.size}{the text size of the legend tick labelings. By
default, it is 5}

\item{legend.title.size}{the text size of the legend titles. By
default, it is 6}

\item{y.text.size}{the text size of the y tick labelings. By default,
it is 6}

\item{x.rotate}{the angle to rotate the x tick labelings. By default,
it is 60}

\item{x.text.size}{the text size of the x tick labelings. By default,
it is 6}

\item{x.text.hjust}{the hjust of the x tick labelings. By default, it
is 0.5}

\item{...}{additional graphic parameters for ggtree::gheatmap}
}
\value{
a ggplot object
}
\description{
\code{xGTheatmapAdv} is supposed to plot tree and heatmap given
associated data on tree tips. It returns an object of class "ggplot".
}
\note{
none
}
\examples{
\dontrun{
# Load the XGR package
library(XGR)
set.seed(825)
tree <- ape::rtree(50)
gp <- xGT(tree)
# heatmap data
x <- matrix(rnorm(length(tree$tip.label)*5), ncol=5)
colnames(x) <- paste0('C',1:5)
rownames(x) <- tree$tip.label
gp_heatmap <- xGTheatmapAdv(gp, x, ratio.width=0.5)
gp_heatmap
gp_heatmap + coord_polar(theta="y")

# advanced usage
gp <- xGT(tree, layout="fan", open.angle=180)
gp_heatmap <- xGTheatmapAdv(gp, x, y.text.size=0)
gp_heatmap
}
}
\seealso{
\code{\link{xGT}}
}
