% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xGTplot.r
\name{xGTplot}
\alias{xGTplot}
\title{Function to draw heatmap on tree tips}
\usage{
xGTplot(gp, data, type = c("bar", "ridge"), combined = T,
colormap = "spectral", ncolors = 64, y.text.size = 2)
}
\arguments{
\item{gp}{a ggplot object resulting from ggtree}

\item{data}{a data frame/matrix for coloring}

\item{type}{the plot type. It can be "bar" and "ridge"}

\item{combined}{logical to indicate whether the tree is combined. By
default, it sets to true for display}

\item{colormap}{short name for the colormap. It can be one of "jet"
(jet colormap), "bwr" (blue-white-red colormap), "gbr" (green-black-red
colormap), "wyr" (white-yellow-red colormap), "br" (black-red
colormap), "yr" (yellow-red colormap), "wb" (white-black colormap), and
"rainbow" (rainbow colormap, that is,
red-yellow-green-cyan-blue-magenta). Alternatively, any
hyphen-separated HTML color names, e.g. "blue-black-yellow",
"royalblue-white-sandybrown", "darkgreen-white-darkviolet". A list of
standard color names can be found in
\url{http://html-color-codes.info/color-names}}

\item{ncolors}{the number of colors specified over the colormap}

\item{y.text.size}{the text size of the y tick labelings. By default,
it is 6}
}
\value{
a ggplot object
}
\description{
\code{xGTplot} is supposed to draw heatmap on tree tips.
}
\note{
none
}
\examples{
\dontrun{
# Load the XGR package
library(XGR)
set.seed(825)
tree <- ape::rtree(50)
gp <- xGGtree(tree)

# barplot plot
data <- data.frame(id=tree$tip.label,
val=rnorm(length(tree$tip.label)))
gp_bar <- xGTplot(gp, data, type="bar")
gp_bar

# ridge plot
id <- rep(tree$tip.label,each=20)
val <- sapply(1:length(tree$tip.label), function(i) rnorm(20,mean=i))
data <- data.frame(id=id, val=as.vector(val))
gp_ridge <- xGTplot(gp, data, type="ridge")
gp_ridge

# combine gp_bar and gp_ridge
g1 <- gtable::ggplotGrob(gp_bar)
g2 <- gtable::ggplotGrob(gp_ridge)
g <- cbind(g1, g2, size = "first")
g$heights <- grid::unit.pmax(g1$heights, g2$heights)
grid::grid.draw(g)
}
}
\seealso{
\code{\link{xGTplot}}
}
