% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xIOcreate.r
\name{xIOcreate}
\alias{xIOcreate}
\title{Function to construct an immune ontology}
\usage{
xIOcreate(cutoff.anno = 15, colormap = "ggplot2", filename =
"xIOcreate",
verbose = TRUE, RData.location =
"http://galahad.well.ox.ac.uk/bigdata")
}
\arguments{
\item{cutoff.anno}{an integer specifying the annotation cutoff obove
which ontology terms are considered}

\item{colormap}{short name for the colormap. It can be one of "jet"
(jet colormap), "bwr" (blue-white-red colormap), "gbr" (green-black-red
colormap), "wyr" (white-yellow-red colormap), "br" (black-red
colormap), "yr" (yellow-red colormap), "wb" (white-black colormap),
"rainbow" (rainbow colormap, that is,
red-yellow-green-cyan-blue-magenta), and "ggplot2" (emulating ggplot2
default color palette). Alternatively, any hyphen-separated HTML color
names, e.g. "lightyellow-orange" (by default), "blue-black-yellow",
"royalblue-white-sandybrown", "darkgreen-white-darkviolet". A list of
standard color names can be found in
\url{http://html-color-codes.info/color-names}}

\item{filename}{the without-extension part of the name of the output
file. By default, it is 'xA2GraphML'}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to true for display}

\item{RData.location}{the characters to tell the location of built-in
RData files. See \code{\link{xRDataLoader}} for details}
}
\value{
a list with 3 components, two ggplot objects (code and table) and an
igraph object (ig) having attributes:
\itemize{
\item{\code{name}: a node attribute for name (ie node.code)}
\item{\code{term_id}: a node attribute for term_id (the primary
source)}
\item{\code{term_name}: a node attribute for term_name (the primary
source)}
\item{\code{term_distance}: a node attribute for term_distance
(starting with 0, 1, ...)}
\item{\code{term_namespace}: a node attribute for term_namespace (ie
IO, Disease, Function, Phenotype)}
\item{\code{anno}: a node attribute as a list storing annotated gene
symbols}
\item{\code{net}: a node attribute as a list storing connectivity
between genes}
\item{\code{path}: a node attribute as a list storing pathways
enriched}
\item{\code{node.label}: a node attribute for the node label in the
form of '[nAnno->nNode@nEdge->nPath] term_name'}
\item{\code{node.label.size}: a node attribute for the node label size
(decreasing radiately)}
\item{\code{node.label.color}: a node attribute for the node label
color (grouped by the subontologies)}
\item{\code{xcoord}: a node attribute for the node x coordinate}
\item{\code{ycoord}: a node attribute for the node y coordinate}
\item{\code{node.code}: a node attribute for the 3-letter node code,
1st letter for the subontology, 2nd for the distance, 3rd for the
sequential order at the same distance}
\item{\code{node.table}: a node attribute for the node label in the
lookup table}
\item{\code{edge.color}: an edge attribute for the edge color (grouped
by the subontologies)}
}
}
\description{
\code{xIOcreate} is supposed to construct an immune ontology.
}
\note{
none
}
\examples{
\dontrun{
# Load the library
library(XGR)
RData.location <- "http://galahad.well.ox.ac.uk/bigdata_dev/"

ls_IO <- xIOcreate(RData.location=RData.location)
ls_IO$ig
ls_IO$code
ls_IO$table

lapply(V(ls_IO$ig)[['b5c']]$path, function(x) xA2GraphML(query=x,
filename=x, RData.location=RData.location))
}
}
\seealso{
\code{\link{xIOcreate}}
}
