% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xMakeGenes.r
\name{xMakeGenes}
\alias{xMakeGenes}
\title{Function to prepare a gene_info RDS file}
\usage{
xMakeGenes(
gene_info.file = NULL,
output.prefix = NULL,
output.dir = "./",
stamp = as.Date(Sys.time())
)
}
\arguments{
\item{gene_info.file}{an input file for gene_info}

\item{output.prefix}{a prefix for the output file}

\item{output.dir}{the output directory}

\item{stamp}{the stamp associated with this RDS file. By default it is
the date when the file created}
}
\value{
an EG object
}
\description{
\code{xMakeGenes} is supposed to prepare a gene_info RDS file.
}
\note{
None
}
\examples{
\dontrun{
EG <- xMakeGenes(gene_info.file="GeneInfo.9606",
output.prefix="org.Hs.eg")
}
}
\seealso{
\code{\link{xMakeGenes}}
}
