% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xMakeGenesets.r
\name{xMakeGenesets}
\alias{xMakeGenesets}
\title{Function to prepare a gene_info RDS file}
\usage{
xMakeGenesets(
association.file = NULL,
set_info.file = NULL,
output.prefix = NULL,
output.dir = "./",
stamp = as.Date(Sys.time())
)
}
\arguments{
\item{association.file}{an input file for association}

\item{set_info.file}{an input file for set_info}

\item{output.prefix}{a prefix for the output file}

\item{output.dir}{the output directory}

\item{stamp}{the stamp associated with this RDS file. By default it is
the date when the file created}
}
\value{
a GS object
}
\description{
\code{xMakeGenesets} is supposed to prepare a gene_info RDS file.
}
\note{
None
}
\examples{
\dontrun{
GS <- xMakeGenesets(association.file="GENE2GOMF.txt",
set_info.file="GO.txt", output.prefix="org.Hs.egGOMF")
}
}
\seealso{
\code{\link{xMakeGenesets}}
}
