% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xMEbesd.r
\name{xMEbesd}
\alias{xMEbesd}
\title{Function to store eQTL summary data in the BESD format}
\usage{
xMEbesd(summary, contexts = c("CD14", "LPS2", "LPS24", "IFN", "Bcell",
"NK", "Neutrophil", "Monocyte", "Blood", "CD4", "CD8", "shared_CD14",
"shared_IFN", "shared_LPS2", "shared_LPS24"), outdir = ".",
clear = T, verbose = T)
}
\arguments{
\item{summary}{a data frame storing eQTL summary statistics. It must
contain columns 'context', and columns
('snp_cse','snps','effect_allele','other_allele','effect_maf') for the
esi file, and columns ('gene_cse','gene','Symbol') for the epi file,
and columns ('snps','gene','beta','statistic','pvalue','FDR') for the
besd file}

\item{contexts}{a vector specifying contexts included}

\item{outdir}{the output directory. By default it is '.'}

\item{clear}{logical to indicate whether the temporary and log files
are cleared up. By default, it sets to TRUE}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to TRUE for display}
}
\value{
a logical vector
}
\description{
\code{xMEbesd} is supposed to store eQTL summary data in the BESD
format ready for the use in SMR.
}
\note{
This function requires the software 'smr' at
\url{http://cnsgenomics.com/software/smr}. Shell command lines in
Terminal (Mac and Linux) are:
\itemize{
\item{1a) Mac: \code{wget
http://cnsgenomics.com/software/smr/download/smr_Mac.zip && unzip
smr_Mac.zip && mv smr_Mac ~/smr}}
\item{1b) Linux: \code{wget
https://cnsgenomics.com/software/smr/download/smr_Linux.zip && unzip
smr_Linux.zip && mv smr_Linux ~/smr}}
\item{2a) # Assuming a ROOT (sudo) privilege: \cr\code{sudo cp ~/smr
/usr/local/bin}}
\item{2b) # Assuming without ROOT (sudo) privilege and adding the
system PATH variable to your ~/.bash_profile file: \cr\code{export
PATH=$HOME:$PATH}}
}
}
\examples{
\dontrun{
# Load the library
library(XGR)
summary <- xRDataLoader('JK_cohort_xMEdb',
RData.location=RData.location)
xMEbesd(summary, contexts="CD14", outdir="Pi_eQTL_hg19")
}
}
\seealso{
\code{\link{xMEbesd}}
}
