% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xPolar.r
\name{xPolar}
\alias{xPolar}
\title{Function to visualise a data frame using a polor plot}
\usage{
xPolar(df, colormap = "darkblue-yellow-darkred", shape = 19, size = 2,
font.family = "sans", signature = TRUE)
}
\arguments{
\item{df}{a data frame with two columns ('name' and 'value')}

\item{colormap}{either NULL or color names ('darkblue-yellow-darkred'
by default) for points}

\item{shape}{an integer specifying point shape}

\item{size}{an integer specifying the point size. By default, it sets
to 2}

\item{font.family}{the font family for texts}

\item{signature}{logical to indicate whether the signature is assigned
to the plot caption. By default, it sets TRUE showing which function is
used to draw this graph}
}
\value{
an object of class "ggplot"
}
\description{
\code{xPolar} is supposed to visualise a data frame using a polor plot.
It returns an object of class "ggplot".
}
\note{
none
}
\seealso{
\code{\link{xPolar}}
}
