% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xPrioritiserManhattan.r
\name{xPrioritiserManhattan}
\alias{xPrioritiserManhattan}
\title{Function to visualise prioritised genes using manhattan plot}
\usage{
xPrioritiserManhattan(pNode, color = c("darkred", "darkgreen"), cex =
0.5,
highlight.top = 10, highlight.col = "deepskyblue",
highlight.label.size = 2, highlight.label.offset = 0.02,
highlight.label.col = "darkblue", verbose = T,
RData.location =
"https://github.com/hfang-bristol/RDataCentre/blob/master/XGR/1.0.0")
}
\arguments{
\item{pNode}{an object of class "pNode"}

\item{color}{a character vector for point colors to alternate}

\item{cex}{a numeric value for point size}

\item{highlight.top}{the number of the top targets to be highlighted}

\item{highlight.col}{the highlight colors}

\item{highlight.label.size}{the highlight label size}

\item{highlight.label.offset}{the highlight label offset}

\item{highlight.label.col}{the highlight label color}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to false for no display}

\item{RData.location}{the characters to tell the location of built-in
RData files. See \code{\link{xRDataLoader}} for details}
}
\value{
an object of class "ggplot"
}
\description{
\code{xPrioritiserManhattan} is supposed to visualise prioritised genes
using manhattan plot.
}
\note{
none
}
\examples{
\dontrun{
# Load the library
library(XGR)
library(igraph)
library(dnet)
library(GenomicRanges)
#library(qqman)
library(ggbio)

RData.location="/Users/hfang/Sites/SVN/github/RDataCentre/XGR/1.0.0"
# a) provide the seed nodes/genes with the weight info
## load ImmunoBase
ImmunoBase <- xRDataLoader(RData.customised='ImmunoBase',
RData.location=RData.location)
## get genes within 500kb away from AS GWAS lead SNPs
seeds.genes <- ImmunoBase$AS$genes_variants
## seeds weighted according to distance away from lead SNPs
data <- 1- seeds.genes/500000

# b) perform priority analysis
pNode <- xPrioritiserGenes(data=data,
network="PCommonsDN_medium",restart=0.7, RData.location=RData.location)

# c) manhattan plot
mp <- xPrioritiserManhattan(pNode, highlight.top=10,
RData.location=RData.location)
#pdf(file="Gene_manhattan.pdf", height=6, width=12, compress=TRUE)
print(mp)
#dev.off()
}
}
\seealso{
\code{\link{xRDataLoader}}, \code{\link{xPrioritiser}},
\code{\link{xPrioritiserSNPs}}, \code{\link{xPrioritiserGenes}},
\code{\link{xPrioritiserPathways}}
}

