% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xPrioritiserPathways.r
\name{xPrioritiserPathways}
\alias{xPrioritiserPathways}
\title{Function to prioritise pathways based on GSEA analysis of prioritised genes}
\usage{
xPrioritiserPathways(pNode, ontology = c("GOBP", "GOMF", "GOCC", "PS",
"PS2",
"SF", "DO", "HPPA", "HPMI", "HPCM", "HPMA", "MP", "MsigdbH",
"MsigdbC1",
"MsigdbC2CGP", "MsigdbC2CPall", "MsigdbC2CP", "MsigdbC2KEGG",
"MsigdbC2REACTOME", "MsigdbC2BIOCARTA", "MsigdbC3TFT", "MsigdbC3MIR",
"MsigdbC4CGN", "MsigdbC4CM", "MsigdbC5BP", "MsigdbC5MF", "MsigdbC5CC",
"MsigdbC6", "MsigdbC7", "DGIdb"), size.range = c(10, 2000),
which.distance = NULL, path.mode = c("all_paths", "shortest_paths",
"all_shortest_paths"), true.path.rule = F, nperm = 1000, verbose = T,
RData.location =
"https://github.com/hfang-bristol/RDataCentre/blob/master/XGR/0.99.0")
}
\arguments{
\item{pNode}{an object of class "pNode"}

\item{ontology}{the ontology supported currently. It can be "GOBP" for
Gene Ontology Biological Process, "GOMF" for Gene Ontology Molecular
Function, "GOCC" for Gene Ontology Cellular Component, "PS" for
phylostratific age information, "PS2" for the collapsed PS version
(inferred ancestors being collapsed into one with the known taxonomy
information), "SF" for domain superfamily assignments, "DO" for Disease
Ontology, "HPPA" for Human Phenotype Phenotypic Abnormality, "HPMI" for
Human Phenotype Mode of Inheritance, "HPCM" for Human Phenotype
Clinical Modifier, "HPMA" for Human Phenotype Mortality Aging, "MP" for
Mammalian Phenotype, and Drug-Gene Interaction database (DGIdb) for
drugable categories, and the molecular signatures database (Msigdb,
including "MsigdbH", "MsigdbC1", "MsigdbC2CGP", "MsigdbC2CPall",
"MsigdbC2CP", "MsigdbC2KEGG", "MsigdbC2REACTOME", "MsigdbC2BIOCARTA",
"MsigdbC3TFT", "MsigdbC3MIR", "MsigdbC4CGN", "MsigdbC4CM",
"MsigdbC5BP", "MsigdbC5MF", "MsigdbC5CC", "MsigdbC6", "MsigdbC7")}

\item{size.range}{the minimum and maximum size of members of each term
in consideration. By default, it sets to a minimum of 10 but no more
than 2000}

\item{which.distance}{which terms with the distance away from the
ontology root (if any) is used to restrict terms in consideration. By
default, it sets to 'NULL' to consider all distances}

\item{path.mode}{the mode of paths induced by vertices/nodes with input
annotation data. It can be "all_paths" for all possible paths to the
root, "shortest_paths" for only one path to the root (for each node in
query), "all_shortest_paths" for all shortest paths to the root (i.e.
for each node, find all shortest paths with the equal lengths)}

\item{true.path.rule}{logical to indicate whether the true-path rule
should be applied to propagate annotations. By default, it sets to
false}

\item{nperm}{the number of random permutations. For each permutation,
gene-score associations will be permutated so that permutation of
gene-term associations is realised}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to false for no display}

\item{RData.location}{the characters to tell the location of built-in
RData files. See \code{\link{xRDataLoader}} for details}
}
\value{
a data frame of nTerm x 10, where nTerm is the number of terms, and the
10 columns:
\itemize{
\item{\code{setID}: Term ID}
\item{\code{setSize}: Number of members in Term}
\item{\code{name}: Term Name}
\item{\code{es}: enrichment score}
\item{\code{nes}: normalised enrichment score. It is the version of
enrichment score but after being normalised by gene set size}
\item{\code{pvalue}: nominal p value}
\item{\code{adjp}: adjusted p value. It is the p value but after being
adjusted for multiple comparisons}
\item{\code{fdr}: false discovery rate (FDR). It is the estimated
probability that the normalised enrichment score represents a false
positive finding}
\item{\code{qvalue}: q value. It is the monotunically increasing FDR}
}
}
\description{
\code{xPrioritiserPathways} is supposed to prioritise pathways given
prioritised genes and the ontology in query. It returns an object of
class "eTerm". It is done via gene set enrichment analysis (GSEA).
}
\note{
none
}
\examples{
\dontrun{
# Load the library
library(XGR)
library(igraph)
library(dnet)
library(GenomicRanges)

RData.location="/Users/hfang/Sites/SVN/github/RDataCentre/XGR/0.99.0"
# a) provide the seed nodes/genes with the weight info
## load ImmunoBase
ImmunoBase <- xRDataLoader(RData.customised='ImmunoBase',
RData.location=RData.location)
## get genes within 500kb away from AS GWAS lead SNPs
seeds.genes <- ImmunoBase$AS$genes_variants
## seeds weighted according to distance away from lead SNPs
data <- 1- seeds.genes/500000

# b) perform priority analysis
pNode <- xPrioritiserGenes(data=data,
network="PCommonsDN_medium",restart=0.7, RData.location=RData.location)

# c) derive pathway-level priority
res <- xPrioritiserPathways(pNode=pNode, ontology="MsigdbC2REACTOME",
RData.location=RData.location)

# d) save enrichment results to the file called 'Pathways_priority.txt'
write.table(res, file="Pathways_priority.txt", sep="\\t",
row.names=FALSE)
}
}
\seealso{
\code{\link{xRDataLoader}}, \code{\link{xPrioritiserSNPs}},
\code{\link{xPrioritiserGenes}}
}

