% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xRdWrap.r
\name{xRdWrap}
\alias{xRdWrap}
\title{Function to wrap texts from Rd files}
\usage{
xRdWrap(path = "./XGR/man", remove.dontrun = FALSE)
}
\arguments{
\item{path}{a directory containing Rd files}

\item{remove.dontrun}{logical to indicate whether to remove the
restriction of not running examples. By default, it sets to FALSE
without any modefications}
}
\value{
none
}
\description{
\code{xRdWrap} is supposed to wrap texts from Rd files under a given
directory.
}
\note{
This auxiliary function helps create a new package. The orignal Rd
files will be replaced with new ones.
}
\examples{
# xRdWrap(path="./XGR/man", remove.dontrun=FALSE)
}
\seealso{
\code{\link{xRdWrap}}
}

