% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xReadGenes.r
\name{xReadGenes}
\alias{xReadGenes}
\title{Function to prepare a gene_info RData file}
\usage{
xReadGenes(gene_info.file = NULL, output.prefix = NULL, output.dir =
"./")
}
\arguments{
\item{gene_info.file}{an input file for gene_info}

\item{output.prefix}{a prefix for the output file}

\item{output.dir}{the output directory}
}
\value{
an EG object
}
\description{
\code{xReadGenes} is supposed to prepare a gene_info RData file.
}
\note{
None
}
\examples{
\dontrun{
# Load the library
library(XGR)

EG <- xReadGenes(gene_info.file="GeneInfo.9606",
output.prefix="org.Hs.eg")
}
}
\seealso{
\code{\link{xReadGenes}}
}
