% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xReadGenesets.r
\name{xReadGenesets}
\alias{xReadGenesets}
\title{Function to prepare a gene_info RData file}
\usage{
xReadGenesets(
association.file = NULL,
set_info.file = NULL,
output.prefix = NULL,
output.dir = "./"
)
}
\arguments{
\item{association.file}{an input file for association}

\item{set_info.file}{an input file for set_info}

\item{output.prefix}{a prefix for the output file}

\item{output.dir}{the output directory}
}
\value{
a GS object
}
\description{
\code{xReadGenesets} is supposed to prepare a gene_info RData file.
}
\note{
None
}
\examples{
\dontrun{
# Load the library
library(XGR)

GS <- xReadGenesets(association.file="GENE2GOMF.txt",
set_info.file="GO.txt", output.prefix="org.Hs.egGOMF")
}
}
\seealso{
\code{\link{xReadGenesets}}
}
