% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xSH.r
\name{xSH}
\alias{xSH}
\title{Function to draw multiple component planes of a supra-hexagonal grid using ggplot2}
\usage{
xSH(sMap, which.components = NULL, customised.comp = NULL,
ncolumns = NULL, colormap = "spectral", ncolors = 64,
zlim = NULL, border.color = "transparent", barwidth = 0.4,
barheight = NULL, nbin = 64, legend.title = "",
legend.text.size = 6, legend.title.size = 8)
}
\arguments{
\item{sMap}{an object of class "sMap"}

\item{which.components}{an integer vector specifying which compopnets
will be visualised. By default, it is NULL meaning all components will
be visualised}

\item{customised.comp}{the customised codebook matrix. It has a high
priority over the built one. It can be a vector}

\item{ncolumns}{an integer specifying the number of columns for a
rectangle grid wherein the component planes are placed. By defaul, it
is NULL (decided on according to the number of component planes that
will be visualised)}

\item{colormap}{short name for the colormap. It can be one of "jet"
(jet colormap), "bwr" (blue-white-red colormap), "gbr" (green-black-red
colormap), "wyr" (white-yellow-red colormap), "br" (black-red
colormap), "yr" (yellow-red colormap), "wb" (white-black colormap), and
"rainbow" (rainbow colormap, that is,
red-yellow-green-cyan-blue-magenta). Alternatively, any
hyphen-separated HTML color names, e.g. "blue-black-yellow",
"royalblue-white-sandybrown", "darkgreen-white-darkviolet". A list of
standard color names can be found in
\url{http://html-color-codes.info/color-names}}

\item{ncolors}{the number of colors specified over the colormap}

\item{zlim}{the minimum and maximum z values for which colors should be
plotted, defaulting to the range of the finite values of displayed
codebook matrix}

\item{border.color}{the border color for each hexagon}

\item{barwidth}{the width of the colorbar. Default value is
'legend.key.width' or 'legend.key.size' in 'theme' or theme}

\item{barheight}{the height of the colorbar. Default value is
'legend.key.height' or 'legend.key.size' in 'theme' or theme}

\item{nbin}{the number of bins for drawing colorbar}

\item{legend.title}{the title of the colorbar. By default, it is ''}

\item{legend.text.size}{the text size of the legend tick labelings}

\item{legend.title.size}{the text size of the legend titles}
}
\value{
a ggplot2 object
}
\description{
\code{xSH} is supposed to draw multiple component planes of a
supra-hexagonal grid using ggplot2.
}
\note{
none
}
\examples{
\dontrun{
library(XGR)

gp <- xSH(sMap)
gp <- xSH(sMap, ncolumns=8)
gp <- xSH(sMap, which.components=1)
gp + theme(legend.position="none")

## advanced use
gp <- xSH(sMap, customised.comp=sMap$codebook[,1:2])

## list of gp
ls_gp <-lapply(1:5, function(i) xSH(sMap, which.components=i) +
theme(legend.position="none"))
}
}
\seealso{
\code{\link{xSH}}
}
