% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xSHbase.r
\name{xSHbase}
\alias{xSHbase}
\title{Function to visualise bases partitioned from a supra-hexagonal grid using ggplot2}
\usage{
xSHbase(sMap, sBase, colormap = "spectral", border.color = "white",
legend.title = "", legend.text.size = 6, legend.title.size = 8)
}
\arguments{
\item{sMap}{an object of class "sMap"}

\item{sBase}{an object of class "sBase". It can be an integer vector
specifying clusters}

\item{colormap}{short name for the colormap. It can be one of "jet"
(jet colormap), "bwr" (blue-white-red colormap), "gbr" (green-black-red
colormap), "wyr" (white-yellow-red colormap), "br" (black-red
colormap), "yr" (yellow-red colormap), "wb" (white-black colormap), and
"rainbow" (rainbow colormap, that is,
red-yellow-green-cyan-blue-magenta). Alternatively, any
hyphen-separated HTML color names, e.g. "blue-black-yellow",
"royalblue-white-sandybrown", "darkgreen-white-darkviolet". A list of
standard color names can be found in
\url{http://html-color-codes.info/color-names}}

\item{border.color}{the border color for each hexagon}

\item{legend.title}{the title of the colorbar. By default, it is ''}

\item{legend.text.size}{the text size of the legend tick labelings. By
default, it is 5}

\item{legend.title.size}{the text size of the legend titles. By
default, it is 6}
}
\value{
a ggplot2 object
}
\description{
\code{xSHbase} is supposed to visualise bases partitioned from a
supra-hexagonal grid using ggplot2.
}
\note{
none
}
\examples{
\dontrun{
library(XGR)

gp <- xSHbase(sMap, sBase)
gp + theme(legend.position="none")

## advanced use
# steps
df <- sMap$polygon \%>\% dplyr::group_by(index) \%>\%
dplyr::summarise(val=unique(stepCentroid))
sBase <- df$val
gp <- xSHbase(sMap, sBase, legend.title='Steps')
# further labelled by hits
df_coord <- data.frame(sMap$coord, hits=sMap$hits, stringsAsFactors=F)
gp + geom_text(data=df_coord, aes(x,y,label=hits))

# angles
df <- sMap$polygon \%>\% dplyr::group_by(index) \%>\%
dplyr::summarise(val=unique(angleCentroid))
sBase <- ceiling(180*(df$val/3.14))
gp <- xSHbase(sMap, sBase, legend.title='Angles')
}
}
\seealso{
\code{\link{xSHbase}}
}
