% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xSNP2nGenes.r
\name{xSNP2nGenes}
\alias{xSNP2nGenes}
\title{Function to define nearby genes given a list of SNPs}
\usage{
xSNP2nGenes(data, distance.max = 2e+05, decay.kernel = c("slow",
"linear",
"rapid"), decay.exponent = 2, GR.SNP = "dbSNP_GWAS",
GR.Gene = "UCSC_genes", verbose = T,
RData.location =
"https://github.com/hfang-bristol/RDataCentre/blob/master/Portal")
}
\arguments{
\item{data}{a input vector containing SNPs}

\item{distance.max}{the maximum distance between genes and SNPs. Only
those genes no far way from this distance will be considered as seed
genes. This parameter will influence the distance-component weights
calculated for nearby SNPs per gene}

\item{decay.kernel}{a character specifying a decay kernel function. It
can be one of 'slow' for slow decay, 'linear' for linear decay, and
'rapid' for rapid decay}

\item{decay.exponent}{a numeric specifying a decay exponent. By
default, it sets to 2}

\item{GR.SNP}{the genomic regions of SNPs. By default, it is
'dbSNP_GWAS', that is, SNPs from dbSNP (version 146) restricted to GWAS
SNPs and their LD SNPs (hg19). It can be 'dbSNP_Common', that is,
Common SNPs from dbSNP (version 146) plus GWAS SNPs and their LD SNPs
(hg19). Alternatively, the user can specify the customised input. To do
so, first save your RData file (containing an GR object) into your
local computer, and make sure the GR object content names refer to
dbSNP IDs. Then, tell "GR.SNP" with your RData file name (with or
without extension), plus specify your file RData path in
"RData.location"}

\item{GR.Gene}{the genomic regions of genes. By default, it is
'UCSC_genes', that is, UCSC known canonical genes (together with
genomic locations) based on human genome assembly hg19. Even the user
can specify the customised input. To do so, first save your RData file
(containing an GR object) into your local computer, and make sure the
GR object content names refer to Gene Symbols. Then, tell "GR.Gene"
with your RData file name (with or without extension), plus specify
your file RData path in "RData.location"}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to true for display}

\item{RData.location}{the characters to tell the location of built-in
RData files. See \code{\link{xRDataLoader}} for details}
}
\value{
a data frame with following columns:
\itemize{
\item{\code{Gene}: nearby genes}
\item{\code{SNP}: SNPs}
\item{\code{Dist}: the genomic distance between the gene and the SNP}
\item{\code{Weight}: the distance weight based on the gnomic distance}
}
}
\description{
\code{xSNP2nGenes} is supposed to define nearby genes given a list of
SNPs within certain distance window. The distance weight is calcualted
as a decaying function of the gene-to-SNP distance.
}
\note{
For details on the decay kernels, please refer to
\code{\link{xVisKernels}}
}
\examples{
\dontrun{
# Load the library
library(Pi)
library(GenomicRanges)

# a) provide the SNPs with the significance info
## get lead SNPs reported in AS GWAS and their significance info (p-values)
AS <- read.delim(file.path(path.package("Pi"),"AS.txt"),
stringsAsFactors=FALSE)

# b) define nearby genes
df_nGenes <- xSNP2nGenes(data=AS$SNP, distance.max=200000,
decay.kernel="slow", decay.exponent=2, RData.location=RData.location)
}
}
\seealso{
\code{\link{xRDataLoader}}, \code{\link{xVisKernels}}
}

