% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xSubneter.r
\name{xSubneter}
\alias{xSubneter}
\title{Function to identify a subnetwork from an input network and the signficance level imposed on its nodes}
\usage{
xSubneter(data, network = c("STRING", "PCOMMONS_UN", "PCOMMONS_DN"),
network.confidence = c("highest", "high", "medium"),
subnet.significance = 0.01, subnet.size = NULL, verbose = T,
RData.location =
"https://github.com/hfang-bristol/RDataCentre/blob/master/XGR/0.99.0")
}
\arguments{
\item{data}{a named input vector containing the sinificance level for
nodes (gene symbols). For this named vector, the element names are gene
symbols, the element values for the significance level (measured as
p-value or fdr). Alternatively, it can be a matrix or data frame with
two columns: 1st column for gene symbols, 2nd column for the
significance level}

\item{network}{the ontology supported currently. It can be "STRING" for
human functional protein association network from the STRING database
(version 10), "PCOMMONS_UN" for human protein interaction undirected
network from Pathway Commons (version 7), "PCOMMONS_DN" for human
protein interaction directed network from Pathway Commons (version 7)}

\item{network.confidence}{the parameter used to control the level of
confidence of network edges that are to be further used. When "network"
is "STRING", network edges with highest confidence are those with
confidence scores=900, high confidence for confidence scores=700,
medium confidence for confidence scores=400. When "network" is either
"PCOMMONS_UN" or "PCOMMONS_DN", network edges with highest confidence
are those supported with the PubMed references plus at least 3
different data sources, high confidence for those supported with the
PubMed references plus at least 2 different data sources, medium
confidence for those supported with at least 2 different data sources}

\item{subnet.significance}{the given significance threshold. By
default, it is set to NULL, meaning there is no constraint on
nodes/genes. If given, those nodes/genes with p-values below this are
considered significant and thus scored positively. Instead, those
p-values above this given significance threshold are considered
insigificant and thus scored negatively}

\item{subnet.size}{the desired number of nodes constrained to the
resulting subnet. It is not nulll, a wide range of significance
thresholds will be scanned to find the optimal significance threshold
leading to the desired number of nodes in the resulting subnet.
Notably, the given significance threshold will be overwritten by this
option}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to true for display}

\item{RData.location}{the characters to tell the location of built-in
RData files. See \code{\link{xRDataLoader}} for details}
}
\value{
a subgraph with a maximum score, an object of class "igraph"
}
\description{
\code{xSubneter} is supposed to identify maximum-scoring subnetwork
from an input graph with the node information on the significance
(measured as p-values or fdr). It returns an object of class "igraph".
}
\note{
The algorithm identifying a subnetwork is implemented in the dnet
package
(http://genomemedicine.biomedcentral.com/articles/10.1186/s13073-014-0064-8).
In brief, from an input network with input node/gene information (the
significant level; p-values or FDR), the way of searching for a
maximum-scoring subnetwork is done as follows. Given the threshold of
tolerable p-value, it gives positive scores for nodes with p-values
below the threshold (nodes of interest), and negative scores for nodes
with threshold-above p-values (intolerable). After score
transformation, the search for a maximum scoring subnetwork is deduced
to find the connected subnetwork that is enriched with positive-score
nodes, allowing for a few negative-score nodes as linkers. This
objective is met through minimum spanning tree finding and
post-processing, previously used as a heuristic solver of
prize-collecting Steiner tree problem. The solver is deterministic,
only determined by the given tolerable p-value threshold. For
identification of the subnetwork with a desired number of nodes, an
iterative procedure is also developed to fine-tune tolerable
thresholds. This explicit control over the node size may be necessary
for guiding follow-up experiments.
}
\examples{
\dontrun{
# Load the library
library(XGR)
library(igraph)
library(dnet)

# a) provide the input nodes/genes with the significance info
## load human genes
org.Hs.eg <- xRDataLoader(RData='org.Hs.eg')
sig <- rbeta(500, shape1=0.5, shape2=1)
data <- data.frame(symbols=org.Hs.eg$gene_info$Symbol[1:500], sig)

# b) perform network analysis
# b1) find maximum-scoring subnet based on the given significance threshold
subnet <- xSubneter(data=data, network="STRING",
subnet.significance=0.01)
# b2) find maximum-scoring subnet with the desired node number=50
subnet <- xSubneter(data=data, network="STRING", subnet.size=50)

# c) save subnet results to the files called 'subnet_edges.txt' and 'subnet_nodes.txt'
output <- igraph::get.data.frame(subnet, what="edges")
utils::write.table(output, file="subnet_edges.txt", sep="\\t",
row.names=FALSE)
output <- igraph::get.data.frame(subnet, what="vertices")
utils::write.table(output, file="subnet_nodes.txt", sep="\\t",
row.names=FALSE)

# d) visualise the identified subnet
## do visualisation
xVisNet(g=subnet, pattern=V(subnet)$score, vertex.shape="sphere")

# e) visualise the identified subnet as a circos plot
library(RCircos)
library(GenomicRanges)
xCircos(g=subnet, entity="Gene")
}
}
\seealso{
\code{\link{xRDataLoader}}
}

