% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xTB2IG.r
\name{xTB2IG}
\alias{xTB2IG}
\title{Function to convert an igraph from one or two tibbles}
\usage{
xTB2IG(edges, nodes = NULL, directed = FALSE, intersected = NULL)
}
\arguments{
\item{edges}{a tibble or data frame for edge attributes}

\item{nodes}{a tibble or data frame for node attributes. It can be
NULL}

\item{directed}{a logic specifying whether to create a directed graph.
By default it is false}

\item{intersected}{a logic or NULL specifying whether to restrict to
only the intersected vertices (from input nodes and edges). By default
it is NULL; it will be forced true (if the input nodes do not contain
all vertices in the input edges); however, the input nodes can contain
vertices not in the input edges. It only works when the input nodes are
not NULL}
}
\value{
an igraph object
}
\description{
\code{xTB2IG} is supposed to convert an igraph from one or two tibbles.
}
\note{
none
}
\examples{
set.seed(825)
ig <- sample_pa(20)
V(ig)$name <- seq(1,vcount(ig))
ig \%>\% xIG2TB('edges') -> edges
ig \%>\% xIG2TB('nodes') -> nodes

xTB2IG(edges, nodes) -> g

# nodes do not cointain all vertices in the edges
nodes[1:10,] -> nodes2
xTB2IG(edges, nodes2) -> g2
}
\seealso{
\code{\link{xTB2IG}}
}
