% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xWCNheatmap.r
\name{xWCNheatmap}
\alias{xWCNheatmap}
\title{Function to draw heatmap from a cModule object}
\usage{
xWCNheatmap(cModule, displayBy = c("tom", "adj"), power_pseudo = NULL,
which.modules = NULL, colormap = "lightgrey-orange-red",
zlim = c(0, 1), filename = "xWCNheatmap", verbose = T)
}
\arguments{
\item{cModule}{an object of class "cModule"}

\item{displayBy}{which matrix used for display. It can be "tom" for the
topological overlap matrix, "adj" for the adjacency matrix}

\item{power_pseudo}{NULL or an integer used to enhance the heatmap
blocks. If NULL, no such enhancement. It can be specified explicitly by
an integer}

\item{which.modules}{an integer vector specifying which modules are
visualised. If NULL (by default), all modules will be used}

\item{colormap}{short name for the colormap. It can be one of "jet"
(jet colormap), "bwr" (blue-white-red colormap), "gbr" (green-black-red
colormap), "wyr" (white-yellow-red colormap), "br" (black-red
colormap), "yr" (yellow-red colormap), "wb" (white-black colormap), and
"rainbow" (rainbow colormap, that is,
red-yellow-green-cyan-blue-magenta). Alternatively, any
hyphen-separated HTML color names, e.g. "blue-black-yellow",
"royalblue-white-sandybrown", "darkgreen-white-darkviolet". A list of
standard color names can be found in
\url{http://html-color-codes.info/color-names}}

\item{zlim}{the minimum and maximum z values for which colors should be
plotted, defaulting to [0,1]}

\item{filename}{the without-extension part of the name of the output
file. By default, it is 'xWCNheatmap'}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to true for display}
}
\value{
a named vector specifying module colors
}
\description{
\code{xWCNheatmap} is supposed to draw heatmap from an object of class
"cModule".
}
\note{
none
}
\examples{
\dontrun{
# Load the library
library(XGR)
RData.location <- "http://galahad.well.ox.ac.uk/bigdata/"

modules_color <- xWCNheatmap(cModule)
modules_color <- xWCNheatmap(cModule, displayBy="tom", power_pseudo=4,
which.modules=c(1,3,10))
modules_color <- xWCNheatmap(cModule, displayBy="adj")
}
}
\seealso{
\code{\link{xWCNheatmap}}
}
