% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xWGCNA.r
\name{xWGCNA}
\alias{xWGCNA}
\title{Function to perform weighted gene correlated network analysis}
\usage{
xWGCNA(data, networkType = c("unsigned", "signed", "signed hybrid"),
powerVector = 1:25, setBeta = "wgcna", RsquaredCut = 0.85,
TOMType = c("signed", "unsigned"), minClusterSize = 30, merge = T,
cutHeight = 0.2, verbose = T)
}
\arguments{
\item{data}{a data matrix or data frame. Rows for genes and columns for
samples}

\item{networkType}{network type. It determines how the adajcency [0,1]
is transformed from correlations ([-1,1]), resulting in a weighted
network adjacency matrix. It can be one of "unsigned" (|cor|^power),
"signed" ((0.5*(1+cor))^power), and "signed hybrid" (cor^power if cor>0
and 0 otherwise)}

\item{powerVector}{a vector of soft thresholding powers for which the
scale-free topology fit indices are to be calculated}

\item{setBeta}{NULL or "wgcna" or an integer. If NULL, beta will be
determined internally; if "wgcna", estimate of an appropriate
soft-thresholding power is the lowest power for which the scale free
topology fit R^2 exceeds "RsquaredCut". It can be specified explicitly
by the beta value (an integer)}

\item{RsquaredCut}{desired minimum scale free topology fitting index
R^2}

\item{TOMType}{TOM type. It determines what will be used as inputs. It
can be one of "unsigned" (taking as inputs the adjacency), "signed"
(taking as inputs the adjacency multiplied by the sign of the
adjacency)}

\item{minClusterSize}{minimum cluster size}

\item{merge}{logical to indicate whether to merge close modules
(measured by the eigengene correlation)}

\item{cutHeight}{maximum dissimilarity (1-cor) that qualifies modules
for merging}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to true for display}
}
\value{
an object of class "cModule", a list with following components:
\itemize{
\item{\code{modules}: a data frame of n X 3 containing module
membership information, where n is the number of nodes, and the 3
columns are "nodes", "modules" (an integer), "rank" (intramodule rank)}
\item{\code{expr}: an input data for nodes assigned to modules}
\item{\code{adj}: a weighted network adjacency matrix}
\item{\code{io}: a named list containing inputs/outputs associated,
including input paramters ("networkType","TOMType","minClusterSize"),
and output results
("beta","r2","fit","num_modules","pattern_modules","gp_fit")}
\item{\code{call}: the call that produced this result}
}
}
\description{
\code{xWGCNA} is supposed to perform weighted gene correlated network
analysis. It returns an object of class "cModule".
}
\note{
none
}
\examples{
\dontrun{
# Load the library
library(XGR)
RData.location <- "http://galahad.well.ox.ac.uk/bigdata/"

cModule <- xWGCNA(data, networkType="unsigned", powerVector=1:25,
setBeta=NULL, RsquaredCut=0.85, TOMType="signed", minClusterSize=30,
merge=T, cutHeight=0.2, verbose=T)
head(cModule$modules)
cModule$io
}
}
\seealso{
\code{\link{xWGCNA}}
}
