% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xWGCNAtom.r
\name{xWGCNAtom}
\alias{xWGCNAtom}
\title{Function to draw TOM plot}
\usage{
xWGCNAtom(cModule, beta_pseudo = NULL, which.modules = NULL,
colormap = "grey-orange-red", zlim = c(0, 1),
filename = "xWGCNAtom", verbose = T)
}
\arguments{
\item{cModule}{an object of class "cModule"}

\item{beta_pseudo}{NULL or an integer used to enhance TOM plot. If
NULL, beta will be the one determined internally. It can be specified
explicitly by the beta value (an integer)}

\item{which.modules}{an integer vector specifying which modules are
visualised. If NULL (by default), all modules will be used}

\item{colormap}{short name for the colormap. It can be one of "jet"
(jet colormap), "bwr" (blue-white-red colormap), "gbr" (green-black-red
colormap), "wyr" (white-yellow-red colormap), "br" (black-red
colormap), "yr" (yellow-red colormap), "wb" (white-black colormap), and
"rainbow" (rainbow colormap, that is,
red-yellow-green-cyan-blue-magenta). Alternatively, any
hyphen-separated HTML color names, e.g. "blue-black-yellow",
"royalblue-white-sandybrown", "darkgreen-white-darkviolet". A list of
standard color names can be found in
\url{http://html-color-codes.info/color-names}}

\item{zlim}{the minimum and maximum z values for which colors should be
plotted, defaulting to [0,1]}

\item{filename}{the without-extension part of the name of the output
file. By default, it is 'xWGCNAtom'}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to true for display}
}
\value{
a named vector specifying module colors
}
\description{
\code{xWGCNAtom} is supposed to draw TOM plot from an object of class
"cModule".
}
\note{
none
}
\examples{
\dontrun{
# Load the library
library(XGR)
RData.location <- "http://galahad.well.ox.ac.uk/bigdata/"

modules_color <- xWGCNAtom(cModule)
modules_color <- xWGCNAtom(cModule, beta_pseudo=4,
which.modules=c(1,3,10))
}
}
\seealso{
\code{\link{xWGCNAtom}}
}
