% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassMethod-XGR.r
\name{cModule}
\alias{cModule}
\alias{print.cModule}
\title{Definition for S3 class \code{cModule}}
\usage{
cModule(mem, data, adj, io)

\method{print}{cModule}(x, ...)
}
\arguments{
\item{mem}{a data frame}

\item{data}{a matrix}

\item{adj}{an adjacency matrix}

\item{io}{a named list}

\item{x}{an object of class \code{cModule}}

\item{...}{other parameters}
}
\value{
an object of S3 class \code{cModule}
}
\description{
\code{cModule} has 4 components: mem, data, adj, io.
}
\examples{
\dontrun{
# Load the library
library(XGR)
}
RData.location <- "http://galahad.well.ox.ac.uk/bigdata"
\dontrun{
cModule(mem, data, adj, io)
}
}
\keyword{S3}
\keyword{classes}
