% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xBigraph.r
\name{xBigraph}
\alias{xBigraph}
\title{Function to obtain communities from a bipartitle graph}
\usage{
xBigraph(g, algorithm = c("louvain", "leading_eigen", "fast_greedy"),
seed = 825, verbose = TRUE)
}
\arguments{
\item{g}{an object of class "igraph" (or "graphNEL") for a bipartitel
graph with a 'type' node attribute}

\item{algorithm}{the algorithm to initialise community from a projected
graph. It can be 'louvain' using multi-level optimization,
'leading_eigen' using leading eigenvector, 'fast_greedy' using greedy
optimization}

\item{seed}{an integer specifying the seed}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to true for display}
}
\value{
It returns an igraph object, appended by node attributes including
"community" for communit memberships, "contribution" for contribution
to its community, "xcoord" for x-coordinates, "ycoord" for
y-coordiates, and by edge attributes including "color" for
between-community edges ('grey70') and within-community edges
('grey95').
}
\description{
\code{xBigraph} is supposed to obtain communities from a bipartitle
graph.
}
\note{
The input graph will has an equal weight if there is no 'weight' edge
attribute associated with
}
\examples{
# 1) generate a random bipartite graph
set.seed(123)
g <- sample_bipartite(100, 50, p=0.1)
V(g)$name <- V(g)

\dontrun{
# 2) obtain its community
ig <- xBigraph(g)
gp <- xA2Net(ig, node.label='name', node.label.size=2,
node.label.color='black', node.label.alpha=0.8, node.label.padding=0,
node.label.arrow=0, node.label.force=0.002, node.shape='type',
node.shape.title='Type', node.xcoord='xcoord', node.ycoord='ycoord',
node.color='community', node.color.title='Community',
colormap='jet.both', ncolors=64, zlim=NULL, node.size='contribution',
node.size.range=c(1,4), node.size.title='Contribution', slim=NULL,
edge.color="color",edge.color.alpha=0.5,edge.curve=0,edge.arrow.gap=0)

## make it discrete for the colorbar
gp +
scale_colour_gradientn(colors=xColormap('jet')(64),limits=c(1,3),breaks=seq(1,3))
+ guides(color=guide_legend(title="Community"))
}
}
\seealso{
\code{\link{xA2Net}}
}
