% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xEnrichBarplot.r
\name{xEnrichBarplot}
\alias{xEnrichBarplot}
\title{Function to visualise enrichment results using a barplot}
\usage{
xEnrichBarplot(eTerm, top_num = 10, displayBy = c("fc", "adjp",
"zscore",
"pvalue"))
}
\arguments{
\item{eTerm}{an object of class "eTerm"}

\item{top_num}{the number of the top terms (sorted according to
'displayBy' below). If it is 'auto', only the signficant terms will be
displayed}

\item{displayBy}{which statistics will be used for displaying. It can
be "adjp" for adjusted p value, "pvalue" for p value, "zscore" for
enrichment z-score, "fc" for enrichment fold change}
}
\value{
an object of class "ggplot"
}
\description{
\code{xEnrichBarplot} is supposed to visualise enrichment results using
a barplot. It returns an object of class "ggplot".
}
\note{
none
}
\examples{
\dontrun{
xEnrichBarplot(eTerm)
}
}
\seealso{
\code{\link{xEnricherGenes}}, \code{\link{xEnricherSNPs}},
\code{\link{xEnrichViewer}}
}

