% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xHeatmap.r
\name{xHeatmap}
\alias{xHeatmap}
\title{Function to draw heatmap using ggplot2}
\usage{
xHeatmap(data, reorder = c("none", "row", "col", "both"), colormap =
"jet",
ncolors = 64, zlim = NULL, barwidth = 0.5, barheight = NULL,
nbin = 64, legend.title = "", x.rotate = 45, x.text.size = 8,
y.text.size = 8, shape = 19, size = 1, font.family = "sans",
na.color = "grey80", data.label = NULL, label.size = 1,
label.color = "black", ...)
}
\arguments{
\item{data}{a data frame/matrix for coloring}

\item{reorder}{how to reorder rows and columns. It can be "none" for no
reordering, "row" for reordering rows according to number of sharings
(by default), "col" for reordering columns, and "both" for reordering
rows and columns}

\item{colormap}{short name for the colormap. It can be one of "jet"
(jet colormap), "bwr" (blue-white-red colormap), "gbr" (green-black-red
colormap), "wyr" (white-yellow-red colormap), "br" (black-red
colormap), "yr" (yellow-red colormap), "wb" (white-black colormap), and
"rainbow" (rainbow colormap, that is,
red-yellow-green-cyan-blue-magenta). Alternatively, any
hyphen-separated HTML color names, e.g. "blue-black-yellow",
"royalblue-white-sandybrown", "darkgreen-white-darkviolet". A list of
standard color names can be found in
\url{http://html-color-codes.info/color-names}}

\item{ncolors}{the number of colors specified over the colormap}

\item{zlim}{the minimum and maximum z values for which colors should be
plotted, defaulting to the range of the finite values of displayed
matrix}

\item{barwidth}{the width of the colorbar. Default value is
'legend.key.width' or 'legend.key.size' in 'theme' or theme}

\item{barheight}{the height of the colorbar. Default value is
'legend.key.height' or 'legend.key.size' in 'theme' or theme}

\item{nbin}{the number of bins for drawing colorbar}

\item{legend.title}{the title of the colorbar. By default, it is ''}

\item{x.rotate}{the angle to rotate the x tick labelings. By default,
it is 45}

\item{x.text.size}{the text size of the x tick labelings. By default,
it is 8}

\item{y.text.size}{the text size of the y tick labelings. By default,
it is 8}

\item{shape}{the number specifying the shape. By default, it is 19}

\item{size}{the number specifying the shape size. By default, it is 1}

\item{font.family}{the font family for texts}

\item{na.color}{the color for NAs. By default, it is 'grey80'}

\item{data.label}{a data frame/matrix used for the labelling}

\item{label.size}{the label size}

\item{label.color}{the label color}

\item{...}{additional graphic parameters for
supraHex::visTreeBootstrap}
}
\value{
If the method is 'gpplot2', it returns a ggplot object, with
components: gene, sample, val, y, x
}
\description{
\code{xHeatmap} is supposed to draw heatmap using ggplot2.
}
\note{
none
}
\examples{
\dontrun{
# Load the XGR package and specify the location of built-in data
library(XGR)
gp <- xHeatmap(data, reorder="none", colormap='darkgreen-lightyellow',
ncolors=64, zlim=c(0,151), x.rotate=45, shape=19, size=3,
x.text.size=6,y.text.size=8)
gp + guides(color=guide_legend())
gp + geom_text(aes(x, y,
label=val),size=1.8,color='red',fontface='bold',na.rm=TRUE,angle=45)
}
}
\seealso{
\code{\link{xHeatmap}}
}
