% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xIG2TB.r
\name{xIG2TB}
\alias{xIG2TB}
\title{Function to convert an igraph into a tibble for nodes or edges}
\usage{
xIG2TB(ig, what = c("edges", "nodes"))
}
\arguments{
\item{ig}{an "igraph" object}

\item{what}{what to extract. It can be "edges" for edges and "nodes"
for nodes}
}
\value{
a tibble object
}
\description{
\code{xIG2TB} is supposed to convert an igraph into a tibble for nodes
or edges.
}
\note{
none
}
\examples{
\dontrun{
# Load the library
library(XGR)
}

RData.location <- "http://galahad.well.ox.ac.uk/bigdata"
\dontrun{
ig <- xDefineNet(network="KEGG", RData.location=RData.location)
ig \%>\% xIG2TB('edges')
ig \%>\% xIG2TB('nodes')
}
}
\seealso{
\code{\link{xTB2IG}}
}
