% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xME.r
\name{xME}
\alias{xME}
\title{Function to perform eQTL analysis using MatrixEQTL}
\usage{
xME(expression, genotype, GR.Gene, GR.SNP, mode = c("cis", "both"),
distance.cis = 1e+06, pvalue.cis = 0.01, pvalue.trans = 1e-05,
which.PCs = NULL, mydir = tempdir(check = T), verbose = T,
silent = F)
}
\arguments{
\item{expression}{a data matrix/frame with genes in rows and samples in
columns}

\item{genotype}{a data (sparse) matrix/frame with SNPs in rows and
samples in columns}

\item{GR.Gene}{an GR object storing the genomic regions of genes}

\item{GR.SNP}{an GR object storing the genomic regions of SNPs}

\item{mode}{a character specifying the eQTL mode. It can be 'cis' for
cis-eQTL analysis, 'both' for both cis-eQTL and trans-eQTL analysis}

\item{distance.cis}{a distance defining cis-eQTL. By default, it is
1Mb}

\item{pvalue.cis}{the p-value threshold outputing the cis-eQTL. By
default, it is 1e-2}

\item{pvalue.trans}{the p-value threshold outputing the trans-eQTL. By
default, it is 1e-5}

\item{which.PCs}{a vector specifying which principle components (PCs)
are used for expression being regressed out. If NULL (by default), no
gression is done}

\item{mydir}{a temporary directory file}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to TRUE for display}

\item{silent}{logical to indicate whether the messages will be silent
completely. By default, it sets to false. If true, verbose will be
forced to be false}
}
\value{
a "MatrixEQTL" object, a list including componets 'cis' and 'trans' (if
mode is 'both'). The 'cis' (or 'trans') contains a data frame 'eqtls'
with following columns
("snps","gene","statistic","pvalue","FDR","beta")
}
\description{
\code{xME} is supposed to perform eQTL analysis using MatrixEQTL.
}
\note{
none
}
\examples{
\dontrun{
# Load the library
library(XGR)
}
}
\seealso{
\code{\link{xRegress}}
}
