% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xSHtree.r
\name{xSHtree}
\alias{xSHtree}
\title{Function to draw multiple component planes reorded within a tree using ggplot2}
\usage{
xSHtree(sMap, phylo = NULL, layout = c("rectangular", "fan",
"partition"), tip.size = 10, embed.width = 1,
colormap = "spectral", zlim = NULL, border.color = "transparent",
barwidth = 0.4, barheight = NULL, nbin = 64, legend.title = "",
legend.text.size = 6, legend.title.size = 8)
}
\arguments{
\item{sMap}{an object of class "sMap"}

\item{phylo}{an object of class "phylo" or NULL}

\item{layout}{the visual layout. It can be "rectangular" (via ggtree)
or circular layout via ggraph including "fan" and "partition"}

\item{tip.size}{the tip/title size}

\item{embed.width}{the width/height of embedded ggplot obejcts}

\item{colormap}{short name for the colormap. It can be one of "jet"
(jet colormap), "bwr" (blue-white-red colormap), "gbr" (green-black-red
colormap), "wyr" (white-yellow-red colormap), "br" (black-red
colormap), "yr" (yellow-red colormap), "wb" (white-black colormap), and
"rainbow" (rainbow colormap, that is,
red-yellow-green-cyan-blue-magenta). Alternatively, any
hyphen-separated HTML color names, e.g. "blue-black-yellow",
"royalblue-white-sandybrown", "darkgreen-white-darkviolet". A list of
standard color names can be found in
\url{http://html-color-codes.info/color-names}}

\item{zlim}{the minimum and maximum z values for which colors should be
plotted, defaulting to the range of the finite values of displayed
codebook matrix}

\item{border.color}{the border color for each hexagon}

\item{barwidth}{the width of the colorbar. Default value is
'legend.key.width' or 'legend.key.size' in 'theme' or theme}

\item{barheight}{the height of the colorbar. Default value is
'legend.key.height' or 'legend.key.size' in 'theme' or theme}

\item{nbin}{the number of bins for drawing colorbar}

\item{legend.title}{the title of the colorbar. By default, it is ''}

\item{legend.text.size}{the text size of the legend tick labelings}

\item{legend.title.size}{the text size of the legend titles}
}
\value{
a ggplot2 object
}
\description{
\code{xSHtree} is supposed to draw multiple component planes reorded
within a tree using ggplot2.
}
\note{
none
}
\examples{
\dontrun{
library(XGR)

gp <- xSHtree(sMap, layout="rectangular", embed.width=1)
gp
ls_gp <- gp$ls_gp

# advanced use
tree <- visTreeBootstrap(t(sMap$codebook))
gp <- xSHtree(sMap, tree, layout="rectangular", embed.width=1)
gp <- xSHtree(sMap, tree, layout="fan", embed.width=0.4)
gp <- xSHtree(sMap, tree, layout="partition", embed.width=0.7)
}
}
\seealso{
\code{\link{xSHtree}}
}
