% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xTB2IG.r
\name{xTB2IG}
\alias{xTB2IG}
\title{Function to convert an igraph from one or two tibbles}
\usage{
xTB2IG(edges, nodes = NULL, directed = F, intersected = NULL)
}
\arguments{
\item{edges}{a tibble or data frame for edge attributes}

\item{nodes}{a tibble or data frame for node attributes. It can be
NULL}

\item{directed}{a logic specifying whether to create a directed graph.
By default it is false}

\item{intersected}{a logic or NULL specifying whether to restrict to
only the intersected vertices (from input nodes and edges). By default
it is NULL; it will be forced true (if the input nodes do not contain
all vertices in the input edges); however, the input nodes can contain
vertices not in the input edges. It only works when the input nodes are
not NULL}
}
\value{
an igraph object
}
\description{
\code{xTB2IG} is supposed to convert an igraph from one or two tibbles.
}
\note{
none
}
\examples{
\dontrun{
# Load the library
library(XGR)
}

RData.location <- "http://galahad.well.ox.ac.uk/bigdata"
\dontrun{
ig <- xDefineNet(network="KEGG", RData.location=RData.location)
ig \%>\% xIG2TB('edges') -> edges
ig \%>\% xIG2TB('nodes') -> nodes

xTB2IG(edges, nodes) -> ig
}
}
\seealso{
\code{\link{xIG2TB}}
}
